@@ -14,9 +14,10 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
-        "illuminate/contracts": "^8.0",
-        "illuminate/macroable": "^8.0"
+        "php": "^8.1",
+        "illuminate/conditionable": "^10.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/macroable": "^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -28,11 +29,11 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "8.x-dev"
+            "dev-master": "10.x-dev"
         }
     },
     "suggest": {
-        "symfony/var-dumper": "Required to use the dump method (^5.4)."
+        "symfony/var-dumper": "Required to use the dump method (^6.2)."
     },
     "config": {
         "sort-packages": true
