@@ -28,8 +28,11 @@ use Webklex\PHPIMAP\Support\Masks\AttachmentMask;
  * @property string type
  * @property string content_type
  * @property string id
+ * @property string hash
  * @property string name
- * @property string disposition
+ * @property string description
+ * @property string filename
+ * @property ?string disposition
  * @property string img_src
  *
  * @method integer getPartNumber()
@@ -42,6 +45,8 @@ use Webklex\PHPIMAP\Support\Masks\AttachmentMask;
  * @method string  setContentType(string $content_type)
  * @method string  getId()
  * @method string  setId(string $id)
+ * @method string  getHash()
+ * @method string  setHash(string $hash)
  * @method string  getSize()
  * @method string  setSize(integer $size)
  * @method string  getName()
@@ -54,33 +59,36 @@ class Attachment {
     /**
      * @var Message $oMessage
      */
-    protected $oMessage;
+    protected Message $oMessage;
 
     /**
      * Used config
      *
      * @var array $config
      */
-    protected $config = [];
+    protected array $config = [];
 
     /** @var Part $part */
-    protected $part;
+    protected Part $part;
 
     /**
      * Attribute holder
      *
      * @var array $attributes
      */
-    protected $attributes = [
-        'content' => null,
-        'type' => null,
-        'part_number' => 0,
+    protected array $attributes = [
+        'content'      => null,
+        'hash'         => null,
+        'type'         => null,
+        'part_number'  => 0,
         'content_type' => null,
-        'id' => null,
-        'name' => null,
-        'disposition' => null,
-        'img_src' => null,
-        'size' => null,
+        'id'           => null,
+        'name'         => null,
+        'filename'     => null,
+        'description'  => null,
+        'disposition'  => null,
+        'img_src'      => null,
+        'size'         => null,
     ];
 
     /**
@@ -88,12 +96,12 @@ class Attachment {
      *
      * @var string $mask
      */
-    protected $mask = AttachmentMask::class;
+    protected string $mask = AttachmentMask::class;
 
     /**
      * Attachment constructor.
-     * @param Message   $oMessage
-     * @param Part      $part
+     * @param Message $oMessage
+     * @param Part $part
      */
     public function __construct(Message $oMessage, Part $part) {
         $this->config = ClientManager::get('options');
@@ -102,9 +110,16 @@ class Attachment {
         $this->part = $part;
         $this->part_number = $part->part_number;
 
-        $default_mask = $this->oMessage->getClient()->getDefaultAttachmentMask();
-        if($default_mask != null) {
-            $this->mask = $default_mask;
+        if ($this->oMessage->getClient()) {
+            $default_mask = $this->oMessage->getClient()?->getDefaultAttachmentMask();
+            if ($default_mask != null) {
+                $this->mask = $default_mask;
+            }
+        } else {
+            $default_mask = ClientManager::getMask("attachment");
+            if ($default_mask != "") {
+                $this->mask = $default_mask;
+            }
         }
 
         $this->findType();
@@ -120,15 +135,15 @@ class Attachment {
      * @throws MethodNotFoundException
      */
     public function __call(string $method, array $arguments) {
-        if(strtolower(substr($method, 0, 3)) === 'get') {
+        if (strtolower(substr($method, 0, 3)) === 'get') {
             $name = Str::snake(substr($method, 3));
 
-            if(isset($this->attributes[$name])) {
+            if (isset($this->attributes[$name])) {
                 return $this->attributes[$name];
             }
 
             return null;
-        }elseif (strtolower(substr($method, 0, 3)) === 'set') {
+        } elseif (strtolower(substr($method, 0, 3)) === 'set') {
             $name = Str::snake(substr($method, 3));
 
             $this->attributes[$name] = array_pop($arguments);
@@ -136,7 +151,7 @@ class Attachment {
             return $this->attributes[$name];
         }
 
-        throw new MethodNotFoundException("Method ".self::class.'::'.$method.'() is not supported');
+        throw new MethodNotFoundException("Method " . self::class . '::' . $method . '() is not supported');
     }
 
     /**
@@ -159,7 +174,7 @@ class Attachment {
      * @return mixed|null
      */
     public function __get($name) {
-        if(isset($this->attributes[$name])) {
+        if (isset($this->attributes[$name])) {
             return $this->attributes[$name];
         }
 
@@ -169,71 +184,83 @@ class Attachment {
     /**
      * Determine the structure type
      */
-    protected function findType() {
-        switch ($this->part->type) {
-            case IMAP::ATTACHMENT_TYPE_MESSAGE:
-                $this->type = 'message';
-                break;
-            case IMAP::ATTACHMENT_TYPE_APPLICATION:
-                $this->type = 'application';
-                break;
-            case IMAP::ATTACHMENT_TYPE_AUDIO:
-                $this->type = 'audio';
-                break;
-            case IMAP::ATTACHMENT_TYPE_IMAGE:
-                $this->type = 'image';
-                break;
-            case IMAP::ATTACHMENT_TYPE_VIDEO:
-                $this->type = 'video';
-                break;
-            case IMAP::ATTACHMENT_TYPE_MODEL:
-                $this->type = 'model';
-                break;
-            case IMAP::ATTACHMENT_TYPE_TEXT:
-                $this->type = 'text';
-                break;
-            case IMAP::ATTACHMENT_TYPE_MULTIPART:
-                $this->type = 'multipart';
-                break;
-            default:
-                $this->type = 'other';
-                break;
-        }
+    protected function findType(): void {
+        $this->type = match ($this->part->type) {
+            IMAP::ATTACHMENT_TYPE_MESSAGE => 'message',
+            IMAP::ATTACHMENT_TYPE_APPLICATION => 'application',
+            IMAP::ATTACHMENT_TYPE_AUDIO => 'audio',
+            IMAP::ATTACHMENT_TYPE_IMAGE => 'image',
+            IMAP::ATTACHMENT_TYPE_VIDEO => 'video',
+            IMAP::ATTACHMENT_TYPE_MODEL => 'model',
+            IMAP::ATTACHMENT_TYPE_TEXT => 'text',
+            IMAP::ATTACHMENT_TYPE_MULTIPART => 'multipart',
+            default => 'other',
+        };
     }
 
     /**
      * Fetch the given attachment
      */
-    protected function fetch() {
-
+    protected function fetch(): void {
         $content = $this->part->content;
 
         $this->content_type = $this->part->content_type;
         $this->content = $this->oMessage->decodeString($content, $this->part->encoding);
 
+        // Create a hash of the raw part - this can be used to identify the attachment in the message context. However,
+        // it is not guaranteed to be unique and collisions are possible.
+        // Some additional online resources:
+        // - https://en.wikipedia.org/wiki/Hash_collision
+        // - https://www.php.net/manual/en/function.hash.php
+        // - https://php.watch/articles/php-hash-benchmark
+        // Benchmark speeds:
+        // -xxh3    ~15.19(GB/s) (requires php-xxhash extension or >= php8.1)
+        // -crc32c  ~14.12(GB/s)
+        // -sha256  ~0.25(GB/s)
+        // xxh3 would be nice to use, because of its extra speed and 32 instead of 8 bytes, but it is not compatible with
+        // php < 8.1. crc32c is the next fastest and is compatible with php >= 5.1. sha256 is the slowest, but is compatible
+        // with php >= 5.1 and is the most likely to be unique. crc32c is the best compromise between speed and uniqueness.
+        // Unique enough for our purposes, but not so slow that it could be a bottleneck.
+        $this->hash = hash("crc32c", $this->part->getHeader()->raw."\r\n\r\n".$this->part->content);
+
         if (($id = $this->part->id) !== null) {
             $this->id = str_replace(['<', '>'], '', $id);
-        }else{
-            $this->id = hash("sha256", (string)microtime(true));
+        }else {
+            $this->id = $this->hash;
         }
 
         $this->size = $this->part->bytes;
         $this->disposition = $this->part->disposition;
 
         if (($filename = $this->part->filename) !== null) {
-            $this->setName($filename);
-        } elseif (($name = $this->part->name) !== null) {
-            $this->setName($name);
-        }else {
-            $this->setName("undefined");
+            $this->filename = $this->decodeName($filename);
+        }
+
+        if (($description = $this->part->description) !== null) {
+            $this->description = $this->part->getHeader()->decode($description);
+        }
+
+        if (($name = $this->part->name) !== null) {
+            $this->name = $this->decodeName($name);
         }
 
         if (IMAP::ATTACHMENT_TYPE_MESSAGE == $this->part->type) {
             if ($this->part->ifdescription) {
-                $this->setName($this->part->description);
-            } else {
-                $this->setName($this->part->subtype);
+                if (!$this->name) {
+                    $this->name = $this->part->description;
+                }
+            } else if (!$this->name) {
+                $this->name = $this->part->subtype;
             }
+        }
+        $this->attributes = array_merge($this->part->getHeader()->getAttributes(), $this->attributes);
+
+        if (!$this->filename) {
+            $this->filename = $this->hash;
+        }
+
+        if (!$this->name && $this->filename != "") {
+            $this->name = $this->filename;
         }
     }
 
@@ -244,25 +271,44 @@ class Attachment {
      *
      * @return boolean
      */
-    public function save(string $path, $filename = null): bool {
-        $filename = $filename ?: $this->getName();
+    public function save(string $path, ?string $filename = null): bool {
+        $filename = $filename ? $this->decodeName($filename) : $this->filename;
 
-        return file_put_contents($path.$filename, $this->getContent()) !== false;
+        return file_put_contents($path . DIRECTORY_SEPARATOR . $filename, $this->getContent()) !== false;
     }
 
     /**
-     * Set the attachment name and try to decode it
-     * @param $name
+     * Decode a given name
+     * @param string|null $name
+     *
+     * @return string
      */
-    public function setName($name) {
-        $decoder = $this->config['decoder']['attachment'];
+    public function decodeName(?string $name): string {
         if ($name !== null) {
-            if($decoder === 'utf-8' && extension_loaded('imap')) {
-                $this->name = \imap_utf8($name);
-            }else{
-                $this->name = mb_decode_mimeheader($name);
+            if (str_contains($name, "''")) {
+                $parts = explode("''", $name);
+                if (EncodingAliases::has($parts[0])) {
+                    $name = implode("''", array_slice($parts, 1));
+                }
             }
+
+            $decoder = $this->config['decoder']['message'];
+            if (preg_match('/=\?([^?]+)\?(Q|B)\?(.+)\?=/i', $name, $matches)) {
+                $name = $this->part->getHeader()->decode($name);
+            } elseif ($decoder === 'utf-8' && extension_loaded('imap')) {
+                $name = \imap_utf8($name);
+            }
+
+            // check if $name is url encoded
+            if (preg_match('/%[0-9A-F]{2}/i', $name)) {
+                $name = urldecode($name);
+            }
+
+            // sanitize $name
+            // order of '..' is important
+            return str_replace(['\\', '/', chr(0), ':', '..'], '', $name);
         }
+        return "";
     }
 
     /**
@@ -270,7 +316,7 @@ class Attachment {
      *
      * @return string|null
      */
-    public function getMimeType(){
+    public function getMimeType(): ?string {
         return (new \finfo())->buffer($this->getContent(), FILEINFO_MIME_TYPE);
     }
 
@@ -279,21 +325,30 @@ class Attachment {
      *
      * @return string|null
      */
-    public function getExtension(){
+    public function getExtension(): ?string {
+        $extension = null;
         $guesser = "\Symfony\Component\Mime\MimeTypes";
         if (class_exists($guesser) !== false) {
             /** @var Symfony\Component\Mime\MimeTypes $guesser */
             $extensions = $guesser::getDefault()->getExtensions($this->getMimeType());
-            return $extensions[0] ?? null;
+            $extension = $extensions[0] ?? null;
         }
-
-        $deprecated_guesser = "\Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser";
-        if (class_exists($deprecated_guesser) !== false){
-            /** @var \Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser $deprecated_guesser */
-            return $deprecated_guesser::getInstance()->guess($this->getMimeType());
+        if ($extension === null) {
+            $deprecated_guesser = "\Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser";
+            if (class_exists($deprecated_guesser) !== false) {
+                /** @var \Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesser $deprecated_guesser */
+                $extension = $deprecated_guesser::getInstance()->guess($this->getMimeType());
+            }
         }
-
-        return null;
+        if ($extension === null) {
+            $parts = explode(".", $this->filename);
+            $extension = count($parts) > 1 ? end($parts) : null;
+        }
+        if ($extension === null) {
+            $parts = explode(".", $this->name);
+            $extension = count($parts) > 1 ? end($parts) : null;
+        }
+        return $extension;
     }
 
     /**
@@ -319,7 +374,7 @@ class Attachment {
      * @return $this
      */
     public function setMask($mask): Attachment {
-        if(class_exists($mask)){
+        if (class_exists($mask)) {
             $this->mask = $mask;
         }
 
@@ -342,12 +397,12 @@ class Attachment {
      * @return mixed
      * @throws MaskNotFoundException
      */
-    public function mask($mask = null){
+    public function mask(string $mask = null): mixed {
         $mask = $mask !== null ? $mask : $this->mask;
-        if(class_exists($mask)){
+        if (class_exists($mask)) {
             return new $mask($this);
         }
 
-        throw new MaskNotFoundException("Unknown mask provided: ".$mask);
+        throw new MaskNotFoundException("Unknown mask provided: " . $mask);
     }
 }
