@@ -15,13 +15,17 @@ namespace Webklex\PHPIMAP;
 use ErrorException;
 use Webklex\PHPIMAP\Connection\Protocols\ImapProtocol;
 use Webklex\PHPIMAP\Connection\Protocols\LegacyProtocol;
-use Webklex\PHPIMAP\Connection\Protocols\Protocol;
 use Webklex\PHPIMAP\Connection\Protocols\ProtocolInterface;
 use Webklex\PHPIMAP\Exceptions\AuthFailedException;
 use Webklex\PHPIMAP\Exceptions\ConnectionFailedException;
+use Webklex\PHPIMAP\Exceptions\EventNotFoundException;
 use Webklex\PHPIMAP\Exceptions\FolderFetchingException;
+use Webklex\PHPIMAP\Exceptions\ImapBadRequestException;
+use Webklex\PHPIMAP\Exceptions\ImapServerErrorException;
 use Webklex\PHPIMAP\Exceptions\MaskNotFoundException;
 use Webklex\PHPIMAP\Exceptions\ProtocolNotSupportedException;
+use Webklex\PHPIMAP\Exceptions\ResponseException;
+use Webklex\PHPIMAP\Exceptions\RuntimeException;
 use Webklex\PHPIMAP\Support\FolderCollection;
 use Webklex\PHPIMAP\Support\Masks\AttachmentMask;
 use Webklex\PHPIMAP\Support\Masks\MessageMask;
@@ -38,30 +42,30 @@ class Client {
     /**
      * Connection resource
      *
-     * @var boolean|Protocol|ProtocolInterface
+     * @var ?ProtocolInterface
      */
-    public $connection = false;
+    public ?ProtocolInterface $connection = null;
 
     /**
      * Server hostname.
      *
      * @var string
      */
-    public $host;
+    public string $host;
 
     /**
      * Server port.
      *
      * @var int
      */
-    public $port;
+    public int $port;
 
     /**
      * Service protocol.
      *
-     * @var int
+     * @var string
      */
-    public $protocol;
+    public string $protocol;
 
     /**
      * Server encryption.
@@ -69,20 +73,20 @@ class Client {
      *
      * @var string
      */
-    public $encryption;
+    public string $encryption;
 
     /**
      * If server has to validate cert.
      *
      * @var bool
      */
-    public $validate_cert = true;
+    public bool $validate_cert = true;
 
     /**
      * Proxy settings
      * @var array
      */
-    protected $proxy = [
+    protected array $proxy = [
         'socket' => null,
         'request_fulluri' => false,
         'username' => null,
@@ -93,63 +97,63 @@ class Client {
      * Connection timeout
      * @var int $timeout
      */
-    public $timeout;
+    public int $timeout;
 
     /**
-     * Account username/
+     * Account username
      *
-     * @var mixed
+     * @var string
      */
-    public $username;
+    public string $username;
 
     /**
      * Account password.
      *
      * @var string
      */
-    public $password;
+    public string $password;
 
     /**
      * Additional data fetched from the server.
      *
-     * @var string
+     * @var array
      */
-    public $extensions;
+    public array $extensions;
 
     /**
      * Account authentication method.
      *
-     * @var string
+     * @var ?string
      */
-    public $authentication;
+    public ?string $authentication;
 
     /**
      * Active folder path.
      *
-     * @var string
+     * @var ?string
      */
-    protected $active_folder = null;
+    protected ?string $active_folder = null;
 
     /**
      * Default message mask
      *
      * @var string $default_message_mask
      */
-    protected $default_message_mask = MessageMask::class;
+    protected string $default_message_mask = MessageMask::class;
 
     /**
      * Default attachment mask
      *
      * @var string $default_attachment_mask
      */
-    protected $default_attachment_mask = AttachmentMask::class;
+    protected string $default_attachment_mask = AttachmentMask::class;
 
     /**
      * Used default account values
      *
      * @var array $default_account_config
      */
-    protected $default_account_config = [
+    protected array $default_account_config = [
         'host' => 'localhost',
         'port' => 993,
         'protocol'  => 'imap',
@@ -182,9 +186,33 @@ class Client {
 
     /**
      * Client destructor
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
      */
     public function __destruct() {
         $this->disconnect();
+    }
+
+    /**
+     * Clone the current Client instance
+     *
+     * @return Client
+     */
+    public function clone(): Client {
+        $client = new self();
+        $client->events = $this->events;
+        $client->timeout = $this->timeout;
+        $client->active_folder = $this->active_folder;
+        $client->default_account_config = $this->default_account_config;
+        $config = $this->getAccountConfig();
+        foreach($config as $key => $value) {
+            $client->setAccountConfig($key, $config, $this->default_account_config);
+        }
+        $client->default_message_mask = $this->default_message_mask;
+        $client->default_attachment_mask = $this->default_message_mask;
+        return $client;
     }
 
     /**
@@ -205,12 +233,25 @@ class Client {
     }
 
     /**
+     * Get the current config
+     *
+     * @return array
+     */
+    public function getConfig(): array {
+        $config = [];
+        foreach($this->default_account_config as $key => $value) {
+            $config[$key] = $this->$key;
+        }
+        return $config;
+    }
+
+    /**
      * Set a specific account config
      * @param string $key
      * @param array $config
      * @param array $default_config
      */
-    private function setAccountConfig(string $key, array $config, array $default_config){
+    private function setAccountConfig(string $key, array $config, array $default_config): void {
         $value = $this->default_account_config[$key];
         if(isset($config[$key])) {
             $value = $config[$key];
@@ -221,10 +262,25 @@ class Client {
     }
 
     /**
+     * Get the current account config
+     *
+     * @return array
+     */
+    public function getAccountConfig(): array {
+        $config = [];
+        foreach($this->default_account_config as $key => $value) {
+            if(property_exists($this, $key)) {
+                $config[$key] = $this->$key;
+            }
+        }
+        return $config;
+    }
+
+    /**
      * Look for a possible events in any available config
      * @param $config
      */
-    protected function setEventsFromConfig($config) {
+    protected function setEventsFromConfig($config): void {
         $this->events = ClientManager::get("events");
         if(isset($config['events'])){
             foreach($config['events'] as $section => $events) {
@@ -239,8 +295,7 @@ class Client {
      *
      * @throws MaskNotFoundException
      */
-    protected function setMaskFromConfig($config) {
-        $default_config  = ClientManager::get("masks");
+    protected function setMaskFromConfig($config): void {
 
         if(isset($config['masks'])){
             if(isset($config['masks']['message'])) {
@@ -250,48 +305,56 @@ class Client {
                     throw new MaskNotFoundException("Unknown mask provided: ".$config['masks']['message']);
                 }
             }else{
-                if(class_exists($default_config['message'])) {
-                    $this->default_message_mask = $default_config['message'];
+                $default_mask  = ClientManager::getMask("message");
+                if($default_mask != ""){
+                    $this->default_message_mask = $default_mask;
                 }else{
-                    throw new MaskNotFoundException("Unknown mask provided: ".$default_config['message']);
+                    throw new MaskNotFoundException("Unknown message mask provided");
                 }
             }
             if(isset($config['masks']['attachment'])) {
                 if(class_exists($config['masks']['attachment'])) {
                     $this->default_attachment_mask = $config['masks']['attachment'];
                 }else{
-                    throw new MaskNotFoundException("Unknown mask provided: ".$config['masks']['attachment']);
+                    throw new MaskNotFoundException("Unknown mask provided: ". $config['masks']['attachment']);
                 }
             }else{
-                if(class_exists($default_config['attachment'])) {
-                    $this->default_attachment_mask = $default_config['attachment'];
+                $default_mask  = ClientManager::getMask("attachment");
+                if($default_mask != ""){
+                    $this->default_attachment_mask = $default_mask;
                 }else{
-                    throw new MaskNotFoundException("Unknown mask provided: ".$default_config['attachment']);
+                    throw new MaskNotFoundException("Unknown attachment mask provided");
                 }
             }
         }else{
-            if(class_exists($default_config['message'])) {
-                $this->default_message_mask = $default_config['message'];
+            $default_mask  = ClientManager::getMask("message");
+            if($default_mask != ""){
+                $this->default_message_mask = $default_mask;
             }else{
-                throw new MaskNotFoundException("Unknown mask provided: ".$default_config['message']);
+                throw new MaskNotFoundException("Unknown message mask provided");
             }
 
-            if(class_exists($default_config['attachment'])) {
-                $this->default_attachment_mask = $default_config['attachment'];
+            $default_mask  = ClientManager::getMask("attachment");
+            if($default_mask != ""){
+                $this->default_attachment_mask = $default_mask;
             }else{
-                throw new MaskNotFoundException("Unknown mask provided: ".$default_config['attachment']);
+                throw new MaskNotFoundException("Unknown attachment mask provided");
             }
         }
-
     }
 
     /**
      * Get the current imap resource
      *
-     * @return bool|Protocol|ProtocolInterface
+     * @return ProtocolInterface
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function getConnection() {
+    public function getConnection(): ProtocolInterface {
         $this->checkConnection();
         return $this->connection;
     }
@@ -307,21 +370,38 @@ class Client {
 
     /**
      * Determine if connection was established and connect if not.
+     * Returns true if the connection was closed and has been reopened.
      *
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function checkConnection() {
-        if (!$this->isConnected()) {
+    public function checkConnection(): bool {
+        try {
+            if (!$this->isConnected()) {
+                $this->connect();
+                return true;
+            }
+        } catch (\Throwable) {
             $this->connect();
         }
+        return false;
     }
 
     /**
      * Force the connection to reconnect
      *
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function reconnect() {
+    public function reconnect(): void {
         if ($this->isConnected()) {
             $this->disconnect();
         }
@@ -333,6 +413,11 @@ class Client {
      *
      * @return $this
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function connect(): Client {
         $this->disconnect();
@@ -347,7 +432,7 @@ class Client {
                 throw new ConnectionFailedException("connection setup failed", 0, new ProtocolNotSupportedException($protocol." is an unsupported protocol"));
             }
             $this->connection = new LegacyProtocol($this->validate_cert, $this->encryption);
-            if (strpos($protocol, "legacy-") === 0) {
+            if (str_starts_with($protocol, "legacy-")) {
                 $protocol = substr($protocol, 7);
             }
             $this->connection->setProtocol($protocol);
@@ -363,9 +448,7 @@ class Client {
 
         try {
             $this->connection->connect($this->host, $this->port);
-        } catch (ErrorException $e) {
-            throw new ConnectionFailedException("connection setup failed", 0, $e);
-        } catch (Exceptions\RuntimeException $e) {
+        } catch (ErrorException|RuntimeException $e) {
             throw new ConnectionFailedException("connection setup failed", 0, $e);
         }
         $this->authenticate();
@@ -376,19 +459,18 @@ class Client {
     /**
      * Authenticate the current session
      *
-     * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws ResponseException
      */
-    protected function authenticate() {
-        try {
-            if ($this->authentication == "oauth") {
-                if (!$this->connection->authenticate($this->username, $this->password)) {
-                    throw new AuthFailedException();
-                }
-            } elseif (!$this->connection->login($this->username, $this->password)) {
+    protected function authenticate(): void {
+        if ($this->authentication == "oauth") {
+            if (!$this->connection->authenticate($this->username, $this->password)->validatedData()) {
                 throw new AuthFailedException();
             }
-        } catch (AuthFailedException $e) {
-            throw new ConnectionFailedException("connection setup failed", 0, $e);
+        } elseif (!$this->connection->login($this->username, $this->password)->validatedData()) {
+            throw new AuthFailedException();
         }
     }
 
@@ -396,9 +478,12 @@ class Client {
      * Disconnect from server.
      *
      * @return $this
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
      */
     public function disconnect(): Client {
-        if ($this->isConnected() && $this->connection !== false) {
+        if ($this->isConnected()) {
             $this->connection->logout();
         }
         $this->active_folder = null;
@@ -409,22 +494,23 @@ class Client {
     /**
      * Get a folder instance by a folder name
      * @param string $folder_name
-     * @param string|bool|null $delimiter
-     *
+     * @param string|null $delimiter
+     * @param bool $utf7
      * @return Folder|null
+     * @throws AuthFailedException
      * @throws ConnectionFailedException
      * @throws FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws ResponseException
+     * @throws RuntimeException
      */
-    public function getFolder(string $folder_name, $delimiter = null) {
-        if ($delimiter !== false && $delimiter !== null) {
-            return $this->getFolderByPath($folder_name);
-        }
-
+    public function getFolder(string $folder_name, ?string $delimiter = null, bool $utf7 = false): ?Folder {
         // Set delimiter to false to force selection via getFolderByName (maybe useful for uncommon folder names)
         $delimiter = is_null($delimiter) ? ClientManager::get('options.delimiter', "/") : $delimiter;
-        if (strpos($folder_name, (string)$delimiter) !== false) {
-            return $this->getFolderByPath($folder_name);
+
+        if (str_contains($folder_name, (string)$delimiter)) {
+            return $this->getFolderByPath($folder_name, $utf7);
         }
 
         return $this->getFolderByName($folder_name);
@@ -433,27 +519,39 @@ class Client {
     /**
      * Get a folder instance by a folder name
      * @param $folder_name
+     * @param bool $soft_fail If true, it will return null instead of throwing an exception
      *
      * @return Folder|null
-     * @throws ConnectionFailedException
      * @throws FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function getFolderByName($folder_name) {
-        return $this->getFolders(false)->where("name", $folder_name)->first();
+    public function getFolderByName($folder_name, bool $soft_fail = false): ?Folder {
+        return $this->getFolders(false, null, $soft_fail)->where("name", $folder_name)->first();
     }
 
     /**
      * Get a folder instance by a folder path
      * @param $folder_path
+     * @param bool $utf7
+     * @param bool $soft_fail If true, it will return null instead of throwing an exception
      *
      * @return Folder|null
+     * @throws AuthFailedException
      * @throws ConnectionFailedException
      * @throws FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws ResponseException
+     * @throws RuntimeException
      */
-    public function getFolderByPath($folder_path) {
-        return $this->getFolders(false)->where("path", $folder_path)->first();
+    public function getFolderByPath($folder_path, bool $utf7 = false, bool $soft_fail = false): ?Folder {
+        if (!$utf7) $folder_path = EncodingAliases::convert($folder_path, "utf-8", "utf7-imap");
+        return $this->getFolders(false, null, $soft_fail)->where("path", $folder_path)->first();
     }
 
     /**
@@ -462,27 +560,32 @@ class Client {
      *
      * @param boolean $hierarchical
      * @param string|null $parent_folder
+     * @param bool $soft_fail If true, it will return an empty collection instead of throwing an exception
      *
      * @return FolderCollection
+     * @throws AuthFailedException
      * @throws ConnectionFailedException
      * @throws FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws ResponseException
+     * @throws RuntimeException
      */
-    public function getFolders(bool $hierarchical = true, string $parent_folder = null): FolderCollection {
+    public function getFolders(bool $hierarchical = true, string $parent_folder = null, bool $soft_fail = false): FolderCollection {
         $this->checkConnection();
         $folders = FolderCollection::make([]);
 
         $pattern = $parent_folder.($hierarchical ? '%' : '*');
-        $items = $this->connection->folders('', $pattern);
+        $items = $this->connection->folders('', $pattern)->validatedData();
 
-        if(is_array($items)){
+        if(!empty($items)){
             foreach ($items as $folder_name => $item) {
                 $folder = new Folder($this, $folder_name, $item["delimiter"], $item["flags"]);
 
                 if ($hierarchical && $folder->hasChildren()) {
                     $pattern = $folder->full_name.$folder->delimiter.'%';
 
-                    $children = $this->getFolders(true, $pattern);
+                    $children = $this->getFolders(true, $pattern, $soft_fail);
                     $folder->setChildren($children);
                 }
 
@@ -490,9 +593,11 @@ class Client {
             }
 
             return $folders;
-        }else{
+        }else if (!$soft_fail){
             throw new FolderFetchingException("failed to fetch any folders");
         }
+
+        return $folders;
     }
 
     /**
@@ -501,27 +606,32 @@ class Client {
      *
      * @param boolean $hierarchical
      * @param string|null $parent_folder
+     * @param bool $soft_fail If true, it will return an empty collection instead of throwing an exception
      *
      * @return FolderCollection
-     * @throws ConnectionFailedException
      * @throws FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function getFoldersWithStatus(bool $hierarchical = true, string $parent_folder = null): FolderCollection {
+    public function getFoldersWithStatus(bool $hierarchical = true, string $parent_folder = null, bool $soft_fail = false): FolderCollection {
         $this->checkConnection();
         $folders = FolderCollection::make([]);
 
         $pattern = $parent_folder.($hierarchical ? '%' : '*');
-        $items = $this->connection->folders('', $pattern);
+        $items = $this->connection->folders('', $pattern)->validatedData();
 
-        if(is_array($items)){
+        if(!empty($items)){
             foreach ($items as $folder_name => $item) {
                 $folder = new Folder($this, $folder_name, $item["delimiter"], $item["flags"]);
 
                 if ($hierarchical && $folder->hasChildren()) {
                     $pattern = $folder->full_name.$folder->delimiter.'%';
 
-                    $children = $this->getFoldersWithStatus(true, $pattern);
+                    $children = $this->getFoldersWithStatus(true, $pattern, $soft_fail);
                     $folder->setChildren($children);
                 }
 
@@ -530,9 +640,11 @@ class Client {
             }
 
             return $folders;
-        }else{
+        }else if (!$soft_fail){
             throw new FolderFetchingException("failed to fetch any folders");
         }
+
+        return $folders;
     }
 
     /**
@@ -540,37 +652,67 @@ class Client {
      * @param string $folder_path
      * @param boolean $force_select
      *
-     * @return array|bool
+     * @return array
      * @throws ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function openFolder(string $folder_path, bool $force_select = false) {
+    public function openFolder(string $folder_path, bool $force_select = false): array {
         if ($this->active_folder == $folder_path && $this->isConnected() && $force_select === false) {
-            return true;
+            return [];
         }
         $this->checkConnection();
         $this->active_folder = $folder_path;
-        return $this->connection->selectFolder($folder_path);
+        return $this->connection->selectFolder($folder_path)->validatedData();
+    }
+
+    /**
+     * Set active folder
+     * @param string|null $folder_path
+     *
+     * @return void
+     */
+    public function setActiveFolder(?string $folder_path = null): void {
+        $this->active_folder = $folder_path;
+    }
+
+    /**
+     * Get active folder
+     *
+     * @return string|null
+     */
+    public function getActiveFolder(): ?string {
+        return $this->active_folder;
     }
 
     /**
      * Create a new Folder
-     * @param string $folder
+     * @param string $folder_path
      * @param boolean $expunge
-     *
+     * @param bool $utf7
      * @return Folder
+     * @throws AuthFailedException
      * @throws ConnectionFailedException
+     * @throws EventNotFoundException
      * @throws FolderFetchingException
-     * @throws Exceptions\EventNotFoundException
-     * @throws Exceptions\RuntimeException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws ResponseException
+     * @throws RuntimeException
      */
-    public function createFolder(string $folder, bool $expunge = true): Folder {
+    public function createFolder(string $folder_path, bool $expunge = true, bool $utf7 = false): Folder {
         $this->checkConnection();
-        $status = $this->connection->createFolder($folder);
+
+        if (!$utf7) $folder_path = EncodingAliases::convert($folder_path, "utf-8", "UTF7-IMAP");
+
+        $status = $this->connection->createFolder($folder_path)->validatedData();
 
         if($expunge) $this->expunge();
 
-        $folder = $this->getFolderByPath($folder);
+        $folder = $this->getFolderByPath($folder_path, true);
         if($status && $folder) {
             $event = $this->getEvent("folder", "new");
             $event::dispatch($folder);
@@ -580,16 +722,51 @@ class Client {
     }
 
     /**
-     * Check a given folder
-     * @param $folder
+     * Delete a given folder
+     * @param string $folder_path
+     * @param boolean $expunge
      *
-     * @return array|bool
+     * @return array
+     * @throws AuthFailedException
      * @throws ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws EventNotFoundException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function checkFolder($folder) {
+    public function deleteFolder(string $folder_path, bool $expunge = true): array {
         $this->checkConnection();
-        return $this->connection->examineFolder($folder);
+
+        $folder = $this->getFolderByPath($folder_path);
+        if ($this->active_folder == $folder->path){
+            $this->active_folder = null;
+        }
+        $status = $this->getConnection()->deleteFolder($folder->path)->validatedData();
+        if ($expunge) $this->expunge();
+
+        $event = $this->getEvent("folder", "deleted");
+        $event::dispatch($folder);
+
+        return $status;
+    }
+
+    /**
+     * Check a given folder
+     * @param string $folder_path
+     *
+     * @return array
+     * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
+     */
+    public function checkFolder(string $folder_path): array {
+        $this->checkConnection();
+        return $this->connection->examineFolder($folder_path)->validatedData();
     }
 
     /**
@@ -597,7 +774,7 @@ class Client {
      *
      * @return string
      */
-    public function getFolderPath(){
+    public function getFolderPath(): string {
         return $this->active_folder;
     }
 
@@ -606,14 +783,18 @@ class Client {
      * Ref.: https://datatracker.ietf.org/doc/html/rfc2971
      *
      * @param array|null $ids
-     * @return array|bool|void|null
+     * @return array
      *
      * @throws ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function Id(array $ids = null) {
+    public function Id(array $ids = null): array {
         $this->checkConnection();
-        return $this->connection->ID($ids);
+        return $this->connection->ID($ids)->validatedData();
     }
 
     /**
@@ -621,11 +802,15 @@ class Client {
      *
      * @return array
      * @throws ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function getQuota(): array {
         $this->checkConnection();
-        return $this->connection->getQuota($this->username);
+        return $this->connection->getQuota($this->username)->validatedData();
     }
 
     /**
@@ -634,32 +819,46 @@ class Client {
      *
      * @return array
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function getQuotaRoot(string $quota_root = 'INBOX'): array {
         $this->checkConnection();
-        return $this->connection->getQuotaRoot($quota_root);
+        return $this->connection->getQuotaRoot($quota_root)->validatedData();
     }
 
     /**
      * Delete all messages marked for deletion
      *
-     * @return bool
+     * @return array
      * @throws ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws AuthFailedException
+     * @throws ResponseException
      */
-    public function expunge(): bool {
+    public function expunge(): array {
         $this->checkConnection();
-        return $this->connection->expunge();
+        return $this->connection->expunge()->validatedData();
     }
 
     /**
      * Set the connection timeout
      * @param integer $timeout
      *
-     * @return Protocol
+     * @return ProtocolInterface
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function setTimeout(int $timeout): Protocol {
+    public function setTimeout(int $timeout): ProtocolInterface {
         $this->timeout = $timeout;
         if ($this->isConnected()) {
             $this->connection->setConnectionTimeout($timeout);
@@ -673,6 +872,11 @@ class Client {
      *
      * @return int
      * @throws ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function getTimeout(): int {
         $this->checkConnection();
