@@ -14,6 +14,7 @@ namespace Webklex\PHPIMAP\Connection\Protocols;
 
 use Webklex\PHPIMAP\ClientManager;
 use Webklex\PHPIMAP\Exceptions\AuthFailedException;
+use Webklex\PHPIMAP\Exceptions\ImapBadRequestException;
 use Webklex\PHPIMAP\Exceptions\MethodNotSupportedException;
 use Webklex\PHPIMAP\Exceptions\RuntimeException;
 use Webklex\PHPIMAP\IMAP;
@@ -25,17 +26,16 @@ use Webklex\PHPIMAP\IMAP;
  */
 class LegacyProtocol extends Protocol {
 
-    protected $protocol = "imap";
-    protected $host = null;
-    protected $port = null;
-    protected $encryption = null;
+    protected string $protocol = "imap";
+    protected string $host = "localhost";
+    protected int $port = 993;
 
     /**
      * Imap constructor.
      * @param bool $cert_validation set to false to skip SSL certificate validation
      * @param mixed $encryption Connection encryption method
      */
-    public function __construct(bool $cert_validation = true, $encryption = false) {
+    public function __construct(bool $cert_validation = true, mixed $encryption = false) {
         $this->setCertValidation($cert_validation);
         $this->encryption = $encryption;
     }
@@ -50,9 +50,9 @@ class LegacyProtocol extends Protocol {
     /**
      * Save the information for a nw connection
      * @param string $host
-     * @param null $port
+     * @param int|null $port
      */
-    public function connect(string $host, $port = null) {
+    public function connect(string $host, int $port = null) {
         if ($this->encryption) {
             $encryption = strtolower($this->encryption);
             if ($encryption == "ssl") {
@@ -69,42 +69,51 @@ class LegacyProtocol extends Protocol {
      * @param string $user username
      * @param string $password password
      *
-     * @return bool
-     * @throws AuthFailedException
-     * @throws RuntimeException
+     * @return Response
      */
-    public function login(string $user, string $password): bool {
-        try {
-            $this->stream = \imap_open(
-                $this->getAddress(),
-                $user,
-                $password,
-                0,
-                $attempts = 3,
-                ClientManager::get('options.open')
-            );
-        } catch (\ErrorException $e) {
-            $errors = \imap_errors();
-            $message = $e->getMessage().'. '.implode("; ", (is_array($errors) ? $errors : array()));
-            throw new AuthFailedException($message);
-        }
-
-        if(!$this->stream) {
-            $errors = \imap_errors();
-            $message = implode("; ", (is_array($errors) ? $errors : array()));
-            throw new AuthFailedException($message);
-        }
-
-        $errors = \imap_errors();
-        if(is_array($errors)) {
-            $status = $this->examineFolder();
-            if($status['exists'] !== 0) {
-                $message = implode("; ", (is_array($errors) ? $errors : array()));
-                throw new RuntimeException($message);
+    public function login(string $user, string $password): Response {
+        return $this->response()->wrap(function($response) use ($user, $password) {
+            /** @var Response $response */
+            try {
+                $this->stream = \imap_open(
+                    $this->getAddress(),
+                    $user,
+                    $password,
+                    0,
+                    $attempts = 3,
+                    ClientManager::get('options.open')
+                );
+                $response->addCommand("imap_open");
+            } catch (\ErrorException $e) {
+                $errors = \imap_errors();
+                $message = $e->getMessage() . '. ' . implode("; ", (is_array($errors) ? $errors : array()));
+                throw new AuthFailedException($message);
             }
-        }
 
-        return $this->stream !== false;
+            if (!$this->stream) {
+                $errors = \imap_errors();
+                $message = implode("; ", (is_array($errors) ? $errors : array()));
+                throw new AuthFailedException($message);
+            }
+
+            $errors = \imap_errors();
+            $response->addCommand("imap_errors");
+            if (is_array($errors)) {
+                $status = $this->examineFolder();
+                $response->stack($status);
+                if ($status->data()['exists'] !== 0) {
+                    $message = implode("; ", $errors);
+                    throw new RuntimeException($message);
+                }
+            }
+
+            if ($this->stream !== false) {
+                return ["TAG" . $response->Noun() . " OK [] Logged in\r\n"];
+            }
+
+            $response->addError("failed to login");
+            return [];
+        });
     }
 
     /**
@@ -112,10 +121,11 @@ class LegacyProtocol extends Protocol {
      * @param string $user username
      * @param string $token access token
      *
-     * @return bool|resource
-     * @throws AuthFailedException|RuntimeException
+     * @return Response
+     * @throws AuthFailedException
+     * @throws RuntimeException
      */
-    public function authenticate(string $user, string $token): bool {
+    public function authenticate(string $user, string $token): Response {
         return $this->login($user, $token);
     }
 
@@ -125,12 +135,12 @@ class LegacyProtocol extends Protocol {
      * @return string
      */
     protected function getAddress(): string {
-        $address = "{".$this->host.":".$this->port."/".$this->protocol;
+        $address = "{" . $this->host . ":" . $this->port . "/" . $this->protocol;
         if (!$this->cert_validation) {
             $address .= '/novalidate-cert';
         }
-        if (in_array($this->encryption,['tls', 'notls', 'ssl'])) {
-            $address .= '/'.$this->encryption;
+        if (in_array($this->encryption, ['tls', 'notls', 'ssl'])) {
+            $address .= '/' . $this->encryption;
         } elseif ($this->encryption === "starttls") {
             $address .= '/tls';
         }
@@ -143,25 +153,25 @@ class LegacyProtocol extends Protocol {
     /**
      * Logout of the current session
      *
-     * @return bool success
+     * @return Response
      */
-    public function logout(): bool {
-        if ($this->stream) {
-            $result = \imap_close($this->stream, IMAP::CL_EXPUNGE);
-            $this->stream = false;
-            $this->uid_cache = null;
-            return $result;
-        }
-        return false;
-    }
-
-    /**
-     * Check if the current session is connected
-     *
-     * @return bool
-     */
-    public function connected(): bool {
-        return boolval($this->stream);
+    public function logout(): Response {
+        return $this->response()->wrap(function($response) {
+            /** @var Response $response */
+            if ($this->stream) {
+                $this->uid_cache = [];
+                $response->addCommand("imap_close");
+                if (\imap_close($this->stream, IMAP::CL_EXPUNGE)) {
+                    $this->stream = false;
+                    return [
+                        0 => "BYE Logging out\r\n",
+                        1 => "TAG" . $response->Noun() . " OK Logout completed (0.001 + 0.000 secs).\r\n",
+                    ];
+                }
+                $this->stream = false;
+            }
+            return [];
+        });
     }
 
     /**
@@ -169,7 +179,7 @@ class LegacyProtocol extends Protocol {
      *
      * @throws MethodNotSupportedException
      */
-    public function getCapabilities(): array {
+    public function getCapabilities(): Response {
         throw new MethodNotSupportedException();
     }
 
@@ -177,10 +187,10 @@ class LegacyProtocol extends Protocol {
      * Change the current folder
      * @param string $folder change to this folder
      *
-     * @return bool|array see examineOrselect()
+     * @return Response see examineOrselect()
      * @throws RuntimeException
      */
-    public function selectFolder(string $folder = 'INBOX') {
+    public function selectFolder(string $folder = 'INBOX'): Response {
         $flags = IMAP::OP_READONLY;
         if (in_array($this->protocol, ["pop3", "nntp"])) {
             $flags = IMAP::NIL;
@@ -189,138 +199,210 @@ class LegacyProtocol extends Protocol {
             throw new RuntimeException("failed to reopen stream.");
         }
 
-        \imap_reopen($this->stream, $this->getAddress().$folder, $flags, 3);
-        $this->uid_cache = null;
-        return $this->examineFolder($folder);
+        return $this->response("imap_reopen")->wrap(function($response) use ($folder, $flags) {
+            /** @var Response $response */
+            \imap_reopen($this->stream, $this->getAddress() . $folder, $flags, 3);
+            $this->uid_cache = [];
+
+            $status = $this->examineFolder($folder);
+            $response->stack($status);
+
+            return $status->data();
+        });
     }
 
     /**
      * Examine a given folder
      * @param string $folder examine this folder
      *
-     * @return bool|array
+     * @return Response
      * @throws RuntimeException
      */
-    public function examineFolder(string $folder = 'INBOX') {
-        if (strpos($folder, ".") === 0) {
+    public function examineFolder(string $folder = 'INBOX'): Response {
+        if (str_starts_with($folder, ".")) {
             throw new RuntimeException("Segmentation fault prevented. Folders starts with an illegal char '.'.");
         }
-        $folder = $this->getAddress().$folder;
-        $status = \imap_status($this->stream, $folder, IMAP::SA_ALL);
-        return [
-            "flags" => [],
-            "exists" => $status->messages,
-            "recent" => $status->recent,
-            "unseen" => $status->unseen,
-            "uidnext" => $status->uidnext,
-        ];
+        return $this->response("imap_status")->wrap(function($response) use ($folder) {
+            /** @var Response $response */
+            $status = \imap_status($this->stream, $this->getAddress() . $folder, IMAP::SA_ALL);
+
+            return $status ? [
+                "flags"   => [],
+                "exists"  => $status->messages,
+                "recent"  => $status->recent,
+                "unseen"  => $status->unseen,
+                "uidnext" => $status->uidnext,
+            ] : [];
+        });
     }
 
     /**
      * Fetch message content
-     * @param array|int $uids
+     * @param int|array $uids
      * @param string $rfc
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array
+     * @return Response
      */
-    public function content($uids, string $rfc = "RFC822", $uid = IMAP::ST_UID): array {
-        $result = [];
-        $uids = is_array($uids) ? $uids : [$uids];
-        foreach ($uids as $id) {
-            $result[$id] = \imap_fetchbody($this->stream, $id, "", $uid ? IMAP::ST_UID : IMAP::NIL);
-        }
-        return $result;
+    public function content(int|array $uids, string $rfc = "RFC822", int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($uids, $uid) {
+            /** @var Response $response */
+
+            $result = [];
+            $uids = is_array($uids) ? $uids : [$uids];
+            foreach ($uids as $id) {
+                $response->addCommand("imap_fetchbody");
+                $result[$id] = \imap_fetchbody($this->stream, $id, "", $uid === IMAP::ST_UID ? IMAP::ST_UID : IMAP::NIL);
+            }
+
+            return $result;
+        });
     }
 
     /**
      * Fetch message headers
-     * @param array|int $uids
+     * @param int|array $uids
      * @param string $rfc
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array
+     * @return Response
      */
-    public function headers($uids, string $rfc = "RFC822", $uid = IMAP::ST_UID): array {
-        $result = [];
-        $uids = is_array($uids) ? $uids : [$uids];
-        foreach ($uids as $id) {
-            $result[$id] = \imap_fetchheader($this->stream, $id, $uid ? IMAP::ST_UID : IMAP::NIL);
-        }
-        return $result;
+    public function headers(int|array $uids, string $rfc = "RFC822", int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($uids, $uid) {
+            /** @var Response $response */
+
+            $result = [];
+            $uids = is_array($uids) ? $uids : [$uids];
+            foreach ($uids as $id) {
+                $response->addCommand("imap_fetchheader");
+                $result[$id] = \imap_fetchheader($this->stream, $id, $uid ? IMAP::ST_UID : IMAP::NIL);
+            }
+
+            return $result;
+        });
     }
 
     /**
      * Fetch message flags
-     * @param array|int $uids
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|array $uids
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array
+     * @return Response
      */
-    public function flags($uids, $uid = IMAP::ST_UID): array {
-        $result = [];
-        $uids = is_array($uids) ? $uids : [$uids];
-        foreach ($uids as $id) {
-            $raw_flags = \imap_fetch_overview($this->stream, $id, $uid ? IMAP::ST_UID : IMAP::NIL);
-            $flags = [];
-            if (is_array($raw_flags) && isset($raw_flags[0])) {
-                $raw_flags = (array) $raw_flags[0];
-                foreach($raw_flags as $flag => $value) {
-                    if ($value === 1 && in_array($flag, ["size", "uid", "msgno", "update"]) === false){
-                        $flags[] = "\\".ucfirst($flag);
+    public function flags(int|array $uids, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($uids, $uid) {
+            /** @var Response $response */
+
+            $result = [];
+            $uids = is_array($uids) ? $uids : [$uids];
+            foreach ($uids as $id) {
+                $response->addCommand("imap_fetch_overview");
+                $raw_flags = \imap_fetch_overview($this->stream, $id, $uid ? IMAP::ST_UID : IMAP::NIL);
+                $flags = [];
+                if (is_array($raw_flags) && isset($raw_flags[0])) {
+                    $raw_flags = (array)$raw_flags[0];
+                    foreach ($raw_flags as $flag => $value) {
+                        if ($value === 1 && in_array($flag, ["size", "uid", "msgno", "update"]) === false) {
+                            $flags[] = "\\" . ucfirst($flag);
+                        }
                     }
                 }
+                $result[$id] = $flags;
             }
-            $result[$uid] = $flags;
-        }
 
-        return $result;
+            return $result;
+        });
+    }
+
+    /**
+     * Fetch message sizes
+     * @param int|array $uids
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     *
+     * @return Response
+     */
+    public function sizes(int|array $uids, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($uids, $uid) {
+            /** @var Response $response */
+            $result = [];
+            $uids = is_array($uids) ? $uids : [$uids];
+            $uid_text = implode("','", $uids);
+            $response->addCommand("imap_fetch_overview");
+            if ($uid == IMAP::ST_UID) {
+                $raw_overview = \imap_fetch_overview($this->stream, $uid_text, IMAP::FT_UID);
+            } else {
+                $raw_overview = \imap_fetch_overview($this->stream, $uid_text);
+            }
+            if ($raw_overview !== false) {
+                foreach ($raw_overview as $overview_element) {
+                    $overview_element = (array)$overview_element;
+                    $result[$overview_element[$uid == IMAP::ST_UID ? 'uid' : 'msgno']] = $overview_element['size'];
+                }
+            }
+            return $result;
+        });
     }
 
     /**
      * Get uid for a given id
      * @param int|null $id message number
      *
-     * @return array|string message number for given message or all messages as array
+     * @return Response message number for given message or all messages as array
      */
-    public function getUid($id = null) {
-        if ($id === null) {
-            if ($this->enable_uid_cache && $this->uid_cache) {
-                return $this->uid_cache;
+    public function getUid(int $id = null): Response {
+        return $this->response()->wrap(function($response) use ($id) {
+            /** @var Response $response */
+            if ($id === null) {
+                if ($this->enable_uid_cache && $this->uid_cache) {
+                    return $this->uid_cache;
+                }
+
+                $overview = $this->overview("1:*");
+                $response->stack($overview);
+                $uids = [];
+                foreach ($overview->data() as $set) {
+                    $uids[$set->msgno] = $set->uid;
+                }
+
+                $this->setUidCache($uids);
+                return $uids;
             }
 
-            $overview = $this->overview("1:*");
-            $uids = [];
-            foreach($overview as $set){
-                $uids[$set->msgno] = $set->uid;
+            $response->addCommand("imap_uid");
+            $uid = \imap_uid($this->stream, $id);
+            if ($uid) {
+                return $uid;
             }
 
-            $this->setUidCache($uids);
-            return $uids;
-        }
-
-        return \imap_uid($this->stream, $id);
+            return [];
+        });
     }
 
     /**
      * Get a message number for a uid
      * @param string $id uid
      *
-     * @return int message number
+     * @return Response message number
      */
-    public function getMessageNumber(string $id): int {
-        return \imap_msgno($this->stream, $id);
+    public function getMessageNumber(string $id): Response {
+        return $this->response("imap_msgno")->wrap(function($response) use ($id) {
+            /** @var Response $response */
+            return \imap_msgno($this->stream, $id);
+        });
     }
 
     /**
      * Get a message overview
      * @param string $sequence uid sequence
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array
+     * @return Response
      */
-    public function overview(string $sequence, $uid = IMAP::ST_UID): array {
-        return \imap_fetch_overview($this->stream, $sequence,$uid ? IMAP::ST_UID : IMAP::NIL);
+    public function overview(string $sequence, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response("imap_fetch_overview")->wrap(function($response) use ($sequence, $uid) {
+            /** @var Response $response */
+            return \imap_fetch_overview($this->stream, $sequence, $uid ? IMAP::ST_UID : IMAP::NIL) ?: [];
+        });
     }
 
     /**
@@ -328,52 +410,65 @@ class LegacyProtocol extends Protocol {
      * @param string $reference mailbox reference for list
      * @param string $folder mailbox name match with wildcards
      *
-     * @return array folders that matched $folder as array(name => array('delimiter' => .., 'flags' => ..))
-     * @throws RuntimeException
+     * @return Response folders that matched $folder as array(name => array('delimiter' => .., 'flags' => ..))
      */
-    public function folders(string $reference = '', string $folder = '*'): array {
-        $result = [];
+    public function folders(string $reference = '', string $folder = '*'): Response {
+        return $this->response("imap_getmailboxes")->wrap(function($response) use ($reference, $folder) {
+            /** @var Response $response */
+            $result = [];
 
-        $items = \imap_getmailboxes($this->stream, $this->getAddress(), $reference.$folder);
-        if(is_array($items)){
-            foreach ($items as $item) {
-                $name = $this->decodeFolderName($item->name);
-                $result[$name] = ['delimiter' => $item->delimiter, 'flags' => []];
+            $items = \imap_getmailboxes($this->stream, $this->getAddress(), $reference . $folder);
+            if (is_array($items)) {
+                foreach ($items as $item) {
+                    $name = $this->decodeFolderName($item->name);
+                    $result[$name] = ['delimiter' => $item->delimiter, 'flags' => []];
+                }
+            } else {
+                throw new RuntimeException(\imap_last_error());
             }
-        }else{
-            throw new RuntimeException(\imap_last_error());
-        }
 
-        return $result;
+            return $result;
+        });
     }
 
     /**
      * Manage flags
-     * @param array $flags flags to set, add or remove - see $mode
+     * @param array|string $flags flags to set, add or remove - see $mode
      * @param int $from message for items or start message if $to !== null
      * @param int|null $to if null only one message ($from) is fetched, else it's the
      *                             last message, INF means last message available
      * @param string|null $mode '+' to add flags, '-' to remove flags, everything else sets the flags as given
      * @param bool $silent if false the return values are the new flags for the wanted messages
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
-     * @param null $item unused attribute
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param string|null $item unused attribute
      *
-     * @return bool|array new flags if $silent is false, else true or false depending on success
+     * @return Response new flags if $silent is false, else true or false depending on success
      */
-    public function store(array $flags, int $from, $to = null, $mode = null, bool $silent = true, $uid = IMAP::ST_UID, $item = null) {
+    public function store(array|string $flags, int $from, int $to = null, string $mode = null, bool $silent = true, int|string $uid = IMAP::ST_UID, string $item = null): Response {
         $flag = trim(is_array($flags) ? implode(" ", $flags) : $flags);
 
-        if ($mode == "+"){
-            $status = \imap_setflag_full($this->stream, $from, $flag, $uid ? IMAP::ST_UID : IMAP::NIL);
-        }else{
-            $status = \imap_clearflag_full($this->stream, $from, $flag, $uid ? IMAP::ST_UID : IMAP::NIL);
-        }
+        return $this->response()->wrap(function($response) use ($mode, $from, $flag, $uid, $silent) {
+            /** @var Response $response */
 
-        if ($silent === true) {
-            return $status;
-        }
+            if ($mode == "+") {
+                $response->addCommand("imap_setflag_full");
+                $status = \imap_setflag_full($this->stream, $from, $flag, $uid ? IMAP::ST_UID : IMAP::NIL);
+            } else {
+                $response->addCommand("imap_clearflag_full");
+                $status = \imap_clearflag_full($this->stream, $from, $flag, $uid ? IMAP::ST_UID : IMAP::NIL);
+            }
 
-        return $this->flags($from);
+            if ($silent === true) {
+                if ($status) {
+                    return [
+                        "TAG" . $response->Noun() . " OK Store completed (0.001 + 0.000 secs).\r\n"
+                    ];
+                }
+                return [];
+            }
+
+            return $this->flags($from);
+        });
     }
 
     /**
@@ -381,19 +476,29 @@ class LegacyProtocol extends Protocol {
      * @param string $folder name of target folder
      * @param string $message full message content
      * @param array|null $flags flags for new message
-     * @param string $date date for new message
+     * @param mixed $date date for new message
      *
-     * @return bool success
+     * @return Response
      */
-    public function appendMessage(string $folder, string $message, $flags = null, $date = null): bool {
-        if ($date != null) {
-            if ($date instanceof \Carbon\Carbon){
-                $date = $date->format('d-M-Y H:i:s O');
+    public function appendMessage(string $folder, string $message, array $flags = null, mixed $date = null): Response {
+        return $this->response("imap_append")->wrap(function($response) use ($folder, $message, $flags, $date) {
+            /** @var Response $response */
+            if ($date != null) {
+                if ($date instanceof \Carbon\Carbon) {
+                    $date = $date->format('d-M-Y H:i:s O');
+                }
+                if (\imap_append($this->stream, $this->getAddress() . $folder, $message, $flags, $date)) {
+                    return [
+                        "OK Append completed (0.001 + 0.000 secs).\r\n"
+                    ];
+                }
+            } else if (\imap_append($this->stream, $this->getAddress() . $folder, $message, $flags)) {
+                return [
+                    "OK Append completed (0.001 + 0.000 secs).\r\n"
+                ];
             }
-            return \imap_append($this->stream, $folder, $message, $flags, $date);
-        }
-
-        return \imap_append($this->stream, $folder, $message, $flags);
+            return [];
+        });
     }
 
     /**
@@ -402,30 +507,48 @@ class LegacyProtocol extends Protocol {
      * @param $from
      * @param int|null $to if null only one message ($from) is fetched, else it's the
      *                         last message, INF means last message available
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return bool success
+     * @return Response
      */
-    public function copyMessage(string $folder, $from, $to = null, $uid = IMAP::ST_UID): bool {
-        return \imap_mail_copy($this->stream, $from, $folder, $uid ? IMAP::ST_UID : IMAP::NIL);
+    public function copyMessage(string $folder, $from, int $to = null, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response("imap_mail_copy")->wrap(function($response) use ($from, $folder, $uid) {
+            /** @var Response $response */
+
+            if (\imap_mail_copy($this->stream, $from, $this->getAddress() . $folder, $uid ? IMAP::ST_UID : IMAP::NIL)) {
+                return [
+                    "TAG" . $response->Noun() . " OK Copy completed (0.001 + 0.000 secs).\r\n"
+                ];
+            }
+            throw new ImapBadRequestException("Invalid ID $from");
+        });
     }
 
     /**
      * Copy multiple messages to the target folder
      * @param array $messages List of message identifiers
      * @param string $folder Destination folder
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array|bool Tokens if operation successful, false if an error occurred
+     * @return Response Tokens if operation successful, false if an error occurred
      */
-    public function copyManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID) {
-        foreach($messages as $msg) {
-            if (!$this->copyMessage($folder, $msg, null, $uid)) {
-                return false;
+    public function copyManyMessages(array $messages, string $folder, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($messages, $folder, $uid) {
+            /** @var Response $response */
+            foreach ($messages as $msg) {
+                $copy_response = $this->copyMessage($folder, $msg, null, $uid);
+                $response->stack($copy_response);
+                if (empty($copy_response->data())) {
+                    return [
+                        "TAG" . $response->Noun() . " BAD Copy failed (0.001 + 0.000 secs).\r\n",
+                        "Invalid ID $msg\r\n"
+                    ];
+                }
             }
-        }
-
-        return $messages;
+            return [
+                "TAG" . $response->Noun() . " OK Copy completed (0.001 + 0.000 secs).\r\n"
+            ];
+        });
     }
 
     /**
@@ -434,30 +557,46 @@ class LegacyProtocol extends Protocol {
      * @param $from
      * @param int|null $to if null only one message ($from) is fetched, else it's the
      *                         last message, INF means last message available
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return bool success
+     * @return Response success
      */
-    public function moveMessage(string $folder, $from, $to = null, $uid = IMAP::ST_UID): bool {
-        return \imap_mail_move($this->stream, $from, $folder, $uid ? IMAP::ST_UID : IMAP::NIL);
+    public function moveMessage(string $folder, $from, int $to = null, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response("imap_mail_move")->wrap(function($response) use ($from, $folder, $uid) {
+            if (\imap_mail_move($this->stream, $from, $this->getAddress() . $folder, $uid ? IMAP::ST_UID : IMAP::NIL)) {
+                return [
+                    "TAG" . $response->Noun() . " OK Move completed (0.001 + 0.000 secs).\r\n"
+                ];
+            }
+            throw new ImapBadRequestException("Invalid ID $from");
+        });
     }
 
     /**
      * Move multiple messages to the target folder
      * @param array $messages List of message identifiers
      * @param string $folder Destination folder
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array|bool Tokens if operation successful, false if an error occurred
+     * @return Response Tokens if operation successful, false if an error occurred
+     * @throws ImapBadRequestException
      */
-    public function moveManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID) {
-        foreach($messages as $msg) {
-            if (!$this->moveMessage($folder, $msg, null, $uid)) {
-                return false;
+    public function moveManyMessages(array $messages, string $folder, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response()->wrap(function($response) use ($messages, $folder, $uid) {
+            foreach ($messages as $msg) {
+                $move_response = $this->moveMessage($folder, $msg, null, $uid);
+                $response = $response->include($response);
+                if (empty($move_response->data())) {
+                    return [
+                        "TAG" . $response->Noun() . " BAD Move failed (0.001 + 0.000 secs).\r\n",
+                        "Invalid ID $msg\r\n"
+                    ];
+                }
             }
-        }
-
-        return $messages;
+            return [
+                "TAG" . $response->Noun() . " OK Move completed (0.001 + 0.000 secs).\r\n"
+            ];
+        });
     }
 
     /**
@@ -465,11 +604,11 @@ class LegacyProtocol extends Protocol {
      * Ref.: https://datatracker.ietf.org/doc/html/rfc2971
      *
      * @param null $ids
-     * @return array|bool|void|null
+     * @return Response
      *
      * @throws MethodNotSupportedException
      */
-    public function ID($ids = null) {
+    public function ID($ids = null): Response {
         throw new MethodNotSupportedException();
     }
 
@@ -477,10 +616,14 @@ class LegacyProtocol extends Protocol {
      * Create a new folder (and parent folders if needed)
      * @param string $folder folder name
      *
-     * @return bool success
+     * @return Response
      */
-    public function createFolder(string $folder): bool {
-        return \imap_createmailbox($this->stream, $folder);
+    public function createFolder(string $folder): Response {
+        return $this->response("imap_createmailbox")->wrap(function($response) use ($folder) {
+            return \imap_createmailbox($this->stream, $this->getAddress() . $folder) ? [
+                0 => "TAG" . $response->Noun() . " OK Create completed (0.004 + 0.000 + 0.003 secs).\r\n",
+            ] : [];
+        });
     }
 
     /**
@@ -488,20 +631,28 @@ class LegacyProtocol extends Protocol {
      * @param string $old old name
      * @param string $new new name
      *
-     * @return bool success
+     * @return Response
      */
-    public function renameFolder(string $old, string $new): bool {
-        return \imap_renamemailbox($this->stream, $old, $new);
+    public function renameFolder(string $old, string $new): Response {
+        return $this->response("imap_renamemailbox")->wrap(function($response) use ($old, $new) {
+            return \imap_renamemailbox($this->stream, $this->getAddress() . $old, $this->getAddress() . $new) ? [
+                0 => "TAG" . $response->Noun() . " OK Move completed (0.004 + 0.000 + 0.003 secs).\r\n",
+            ] : [];
+        });
     }
 
     /**
      * Delete a folder
      * @param string $folder folder name
      *
-     * @return bool success
+     * @return Response
      */
-    public function deleteFolder(string $folder): bool {
-        return \imap_deletemailbox($this->stream, $folder);
+    public function deleteFolder(string $folder): Response {
+        return $this->response("imap_deletemailbox")->wrap(function($response) use ($folder) {
+            return \imap_deletemailbox($this->stream, $this->getAddress() . $folder) ? [
+                0 => "OK Delete completed (0.004 + 0.000 + 0.003 secs).\r\n",
+            ] : [];
+        });
     }
 
     /**
@@ -510,7 +661,7 @@ class LegacyProtocol extends Protocol {
      *
      * @throws MethodNotSupportedException
      */
-    public function subscribeFolder(string $folder): bool {
+    public function subscribeFolder(string $folder): Response {
         throw new MethodNotSupportedException();
     }
 
@@ -520,17 +671,21 @@ class LegacyProtocol extends Protocol {
      *
      * @throws MethodNotSupportedException
      */
-    public function unsubscribeFolder(string $folder): bool {
+    public function unsubscribeFolder(string $folder): Response {
         throw new MethodNotSupportedException();
     }
 
     /**
      * Apply session saved changes to the server
      *
-     * @return bool success
+     * @return Response
      */
-    public function expunge(): bool {
-        return \imap_expunge($this->stream);
+    public function expunge(): Response {
+        return $this->response("imap_expunge")->wrap(function($response) {
+            return \imap_expunge($this->stream) ? [
+                0 => "TAG" . $response->Noun() . " OK Expunge completed (0.001 + 0.000 secs).\r\n",
+            ] : [];
+        });
     }
 
     /**
@@ -538,7 +693,7 @@ class LegacyProtocol extends Protocol {
      *
      * @throws MethodNotSupportedException
      */
-    public function noop(): bool {
+    public function noop(): Response {
         throw new MethodNotSupportedException();
     }
 
@@ -563,29 +718,29 @@ class LegacyProtocol extends Protocol {
     /**
      * Search for matching messages
      * @param array $params
-     * @param int $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
+     * @param int|string $uid set to IMAP::ST_UID if you pass message unique identifiers instead of numbers.
      *
-     * @return array message ids
+     * @return Response message ids
      */
-    public function search(array $params, $uid = IMAP::ST_UID): array {
-        $result = \imap_search($this->stream, $params[0], $uid ? IMAP::ST_UID : IMAP::NIL);
-        if ($result === false) {
-            return [];
-        }
-        return $result;
+    public function search(array $params, int|string $uid = IMAP::ST_UID): Response {
+        return $this->response("imap_search")->wrap(function($response) use ($params, $uid) {
+            $response->setCanBeEmpty(true);
+            $result = \imap_search($this->stream, $params[0], $uid ? IMAP::ST_UID : IMAP::NIL);
+            return $result ?: [];
+        });
     }
 
     /**
      * Enable the debug mode
      */
-    public function enableDebug(){
+    public function enableDebug() {
         $this->debug = true;
     }
 
     /**
      * Disable the debug mode
      */
-    public function disableDebug(){
+    public function disableDebug() {
         $this->debug = false;
     }
 
@@ -597,8 +752,8 @@ class LegacyProtocol extends Protocol {
      *
      * @return array|false|string|string[]|null
      */
-    protected function decodeFolderName($name) {
-        preg_match('#\{(.*)\}(.*)#', $name, $preg);
+    protected function decodeFolderName($name): array|bool|string|null {
+        preg_match('#\{(.*)}(.*)#', $name, $preg);
         return mb_convert_encoding($preg[2], "UTF-8", "UTF7-IMAP");
     }
 
@@ -613,20 +768,26 @@ class LegacyProtocol extends Protocol {
      * Retrieve the quota level settings, and usage statics per mailbox
      * @param $username
      *
-     * @return array
+     * @return Response
      */
-    public function getQuota($username): array {
-        return \imap_get_quota($this->stream, 'user.'.$username);
+    public function getQuota($username): Response {
+        return $this->response("imap_get_quota")->wrap(function($response) use ($username) {
+            $result = \imap_get_quota($this->stream, 'user.' . $username);
+            return $result ?: [];
+        });
     }
 
     /**
      * Retrieve the quota settings per user
      * @param string $quota_root
      *
-     * @return array
+     * @return Response
      */
-    public function getQuotaRoot(string $quota_root = 'INBOX'): array {
-        return \imap_get_quotaroot($this->stream, $quota_root);
+    public function getQuotaRoot(string $quota_root = 'INBOX'): Response {
+        return $this->response("imap_get_quotaroot")->wrap(function($response) use ($quota_root) {
+            $result = \imap_get_quotaroot($this->stream, $this->getAddress() . $quota_root);
+            return $result ?: [];
+        });
     }
 
     /**
@@ -639,5 +800,15 @@ class LegacyProtocol extends Protocol {
         }
         $this->protocol = $protocol;
         return $this;
+    }
+
+    /**
+     * Create a new Response instance
+     * @param string|null $command
+     *
+     * @return Response
+     */
+    protected function response(?string $command = ""): Response {
+        return Response::make(0, $command == "" ? [] : [$command], [], $this->debug);
     }
 }
