@@ -22,16 +22,18 @@ use Webklex\PHPIMAP\Message;
 class MessageMovedEvent extends Event {
 
     /** @var Message $old_message */
-    public $old_message;
+    public Message $old_message;
+
     /** @var Message $new_message */
-    public $new_message;
+    public Message $new_message;
 
     /**
      * Create a new event instance.
      * @var Message[] $messages
+     *
      * @return void
      */
-    public function __construct($messages) {
+    public function __construct(array $messages) {
         $this->old_message = $messages[0];
         $this->new_message = $messages[1];
     }
