@@ -14,12 +14,23 @@ namespace Webklex\PHPIMAP;
 
 use ReflectionClass;
 use ReflectionException;
+use Webklex\PHPIMAP\Exceptions\AuthFailedException;
+use Webklex\PHPIMAP\Exceptions\ConnectionFailedException;
+use Webklex\PHPIMAP\Exceptions\EventNotFoundException;
+use Webklex\PHPIMAP\Exceptions\FolderFetchingException;
+use Webklex\PHPIMAP\Exceptions\GetMessagesFailedException;
+use Webklex\PHPIMAP\Exceptions\ImapBadRequestException;
+use Webklex\PHPIMAP\Exceptions\ImapServerErrorException;
 use Webklex\PHPIMAP\Exceptions\InvalidMessageDateException;
 use Webklex\PHPIMAP\Exceptions\MaskNotFoundException;
 use Webklex\PHPIMAP\Exceptions\MessageContentFetchingException;
 use Webklex\PHPIMAP\Exceptions\MessageFlagException;
 use Webklex\PHPIMAP\Exceptions\MessageHeaderFetchingException;
+use Webklex\PHPIMAP\Exceptions\MessageNotFoundException;
+use Webklex\PHPIMAP\Exceptions\MessageSizeFetchingException;
 use Webklex\PHPIMAP\Exceptions\MethodNotFoundException;
+use Webklex\PHPIMAP\Exceptions\ResponseException;
+use Webklex\PHPIMAP\Exceptions\RuntimeException;
 use Webklex\PHPIMAP\Support\AttachmentCollection;
 use Webklex\PHPIMAP\Support\FlagCollection;
 use Webklex\PHPIMAP\Support\Masks\MessageMask;
@@ -35,6 +46,7 @@ use Webklex\PHPIMAP\Traits\HasEvents;
  * @property integer msglist
  * @property integer uid
  * @property integer msgn
+ * @property integer size
  * @property Attribute subject
  * @property Attribute message_id
  * @property Attribute message_no
@@ -52,6 +64,7 @@ use Webklex\PHPIMAP\Traits\HasEvents;
  * @method integer setMsglist($msglist)
  * @method integer getUid()
  * @method integer getMsgn()
+ * @method integer getSize()
  * @method Attribute getPriority()
  * @method Attribute getSubject()
  * @method Attribute getMessageId()
@@ -72,102 +85,102 @@ class Message {
     /**
      * Client instance
      *
-     * @var Client
+     * @var ?Client
      */
-    private $client = Client::class;
+    private ?Client $client = null;
 
     /**
      * Default mask
      *
      * @var string $mask
      */
-    protected $mask = MessageMask::class;
+    protected string $mask = MessageMask::class;
 
     /**
      * Used config
      *
      * @var array $config
      */
-    protected $config = [];
+    protected array $config = [];
 
     /**
      * Attribute holder
      *
      * @var Attribute[]|array $attributes
      */
-    protected $attributes = [];
+    protected array $attributes = [];
 
     /**
      * The message folder path
      *
      * @var string $folder_path
      */
-    protected $folder_path;
+    protected string $folder_path;
 
     /**
      * Fetch body options
      *
-     * @var integer
+     * @var ?integer
      */
-    public $fetch_options = null;
+    public ?int $fetch_options = null;
 
     /**
      * @var integer
      */
-    protected $sequence = IMAP::NIL;
+    protected int $sequence = IMAP::NIL;
 
     /**
      * Fetch body options
      *
      * @var bool
      */
-    public $fetch_body = null;
+    public bool $fetch_body = true;
 
     /**
      * Fetch flags options
      *
      * @var bool
      */
-    public $fetch_flags = null;
+    public bool $fetch_flags = true;
 
     /**
-     * @var Header $header
+     * @var ?Header $header
      */
-    public $header = null;
+    public ?Header $header = null;
 
     /**
      * Raw message body
      *
-     * @var null|string $raw_body
+     * @var string $raw_body
      */
-    public $raw_body = null;
+    protected string $raw_body = "";
 
     /**
      * Message structure
      *
-     * @var Structure $structure
+     * @var ?Structure $structure
      */
-    protected $structure = null;
+    protected ?Structure $structure = null;
 
     /**
      * Message body components
      *
-     * @var array   $bodies
+     * @var array $bodies
      */
-    public $bodies = [];
+    public array $bodies = [];
 
     /** @var AttachmentCollection $attachments */
-    public $attachments;
+    public AttachmentCollection $attachments;
 
     /** @var FlagCollection $flags */
-    public $flags;
+    public FlagCollection $flags;
 
     /**
      * A list of all available and supported flags
      *
-     * @var array $available_flags
+     * @var ?array $available_flags
      */
-    private $available_flags = null;
+    private ?array $available_flags = null;
 
     /**
      * Message constructor.
@@ -179,20 +192,23 @@ class Message {
      * @param boolean $fetch_flags
      * @param integer|null $sequence
      *
-     * @throws Exceptions\ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
-     * @throws Exceptions\RuntimeException
-     * @throws MessageHeaderFetchingException
      * @throws MessageContentFetchingException
-     * @throws Exceptions\EventNotFoundException
      * @throws MessageFlagException
-     * @throws Exceptions\MessageNotFoundException
+     * @throws MessageHeaderFetchingException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function __construct(int $uid, $msglist, Client $client, int $fetch_options = null, bool $fetch_body = false, bool $fetch_flags = false, int $sequence = null) {
+    public function __construct(int $uid, ?int $msglist, Client $client, int $fetch_options = null, bool $fetch_body = false, bool $fetch_flags = false, int $sequence = null) {
         $this->boot();
 
         $default_mask = $client->getDefaultMessageMask();
-        if($default_mask != null) {
+        if ($default_mask != null) {
             $this->mask = $default_mask;
         }
         $this->events["message"] = $client->getDefaultEvents("message");
@@ -237,29 +253,32 @@ class Message {
      * @param null $sequence
      *
      * @return Message
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
-     * @throws ReflectionException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
-     * @throws Exceptions\MessageNotFoundException
+     * @throws ReflectionException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public static function make(int $uid, $msglist, Client $client, string $raw_header, string $raw_body, array $raw_flags, $fetch_options = null, $sequence = null): Message {
+    public static function make(int $uid, ?int $msglist, Client $client, string $raw_header, string $raw_body, array $raw_flags, $fetch_options = null, $sequence = null): Message {
         $reflection = new ReflectionClass(self::class);
-        /** @var self $instance */
+        /** @var Message $instance */
         $instance = $reflection->newInstanceWithoutConstructor();
         $instance->boot();
 
         $default_mask = $client->getDefaultMessageMask();
-        if($default_mask != null) {
+        if ($default_mask != null) {
             $instance->setMask($default_mask);
         }
         $instance->setEvents([
-            "message" => $client->getDefaultEvents("message"),
-            "flag" => $client->getDefaultEvents("flag"),
-        ]);
+                                 "message" => $client->getDefaultEvents("message"),
+                                 "flag"    => $client->getDefaultEvents("flag"),
+                             ]);
         $instance->setFolderPath($client->getFolderPath());
         $instance->setSequence($sequence);
         $instance->setFetchOption($fetch_options);
@@ -276,9 +295,74 @@ class Message {
     }
 
     /**
+     * Create a new message instance by reading and loading a file or remote location
+     *
+     * @throws RuntimeException
+     * @throws MessageContentFetchingException
+     * @throws ResponseException
+     * @throws ImapBadRequestException
+     * @throws InvalidMessageDateException
+     * @throws ConnectionFailedException
+     * @throws ImapServerErrorException
+     * @throws ReflectionException
+     * @throws AuthFailedException
+     * @throws MaskNotFoundException
+     */
+    public static function fromFile($filename): Message {
+        $blob = file_get_contents($filename);
+        if ($blob === false) {
+            throw new RuntimeException("Unable to read file");
+        }
+        return self::fromString($blob);
+    }
+
+    /**
+     * Create a new message instance by reading and loading a string
+     * @param string $blob
+     *
+     * @return Message
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws InvalidMessageDateException
+     * @throws MaskNotFoundException
+     * @throws MessageContentFetchingException
+     * @throws ReflectionException
+     * @throws ResponseException
+     * @throws RuntimeException
+     */
+    public static function fromString(string $blob): Message {
+        $reflection = new ReflectionClass(self::class);
+        /** @var Message $instance */
+        $instance = $reflection->newInstanceWithoutConstructor();
+        $instance->boot();
+
+        $default_mask  = ClientManager::getMask("message");
+        if($default_mask != ""){
+            $instance->setMask($default_mask);
+        }else{
+            throw new MaskNotFoundException("Unknown message mask provided");
+        }
+
+        if(!str_contains($blob, "\r\n")){
+            $blob = str_replace("\n", "\r\n", $blob);
+        }
+        $raw_header = substr($blob, 0, strpos($blob, "\r\n\r\n"));
+        $raw_body = substr($blob, strlen($raw_header)+4);
+
+        $instance->parseRawHeader($raw_header);
+        $instance->parseRawBody($raw_body);
+
+        $instance->setUid(0);
+
+        return $instance;
+    }
+
+    /**
      * Boot a new instance
      */
-    public function boot(){
+    public function boot(): void {
         $this->attributes = [];
 
         $this->config = ClientManager::get('options');
@@ -294,22 +378,30 @@ class Message {
      * @param array $arguments
      *
      * @return mixed
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws MessageNotFoundException
      * @throws MethodNotFoundException
+     * @throws MessageSizeFetchingException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function __call(string $method, array $arguments) {
-        if(strtolower(substr($method, 0, 3)) === 'get') {
+        if (strtolower(substr($method, 0, 3)) === 'get') {
             $name = Str::snake(substr($method, 3));
             return $this->get($name);
-        }elseif (strtolower(substr($method, 0, 3)) === 'set') {
+        } elseif (strtolower(substr($method, 0, 3)) === 'set') {
             $name = Str::snake(substr($method, 3));
 
-            if(in_array($name, array_keys($this->attributes))) {
+            if (in_array($name, array_keys($this->attributes))) {
                 return $this->__set($name, array_pop($arguments));
             }
 
         }
 
-        throw new MethodNotFoundException("Method ".self::class.'::'.$method.'() is not supported');
+        throw new MethodNotFoundException("Method " . self::class . '::' . $method . '() is not supported');
     }
 
     /**
@@ -330,6 +422,14 @@ class Message {
      * @param $name
      *
      * @return Attribute|mixed|null
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws MessageNotFoundException
+     * @throws MessageSizeFetchingException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function __get($name) {
         return $this->get($name);
@@ -340,10 +440,32 @@ class Message {
      * @param $name
      *
      * @return Attribute|mixed|null
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws MessageNotFoundException
+     * @throws RuntimeException
+     * @throws ResponseException
+     * @throws MessageSizeFetchingException
      */
-    public function get($name) {
-        if(isset($this->attributes[$name])) {
+    public function get($name): mixed {
+        if (isset($this->attributes[$name]) && $this->attributes[$name] !== null) {
             return $this->attributes[$name];
+        }
+
+        switch ($name){
+            case "uid":
+                $this->attributes[$name] = $this->client->getConnection()->getUid($this->msgn)->validate()->integer();
+                return $this->attributes[$name];
+            case "msgn":
+                $this->attributes[$name] = $this->client->getConnection()->getMessageNumber($this->uid)->validate()->integer();
+                return $this->attributes[$name];
+            case "size":
+                if (!isset($this->attributes[$name])) {
+                    $this->fetchSize();
+                }
+                return $this->attributes[$name];
         }
 
         return $this->header->get($name);
@@ -355,17 +477,17 @@ class Message {
      * @return bool
      */
     public function hasTextBody(): bool {
-        return isset($this->bodies['text']);
+        return isset($this->bodies['text']) && $this->bodies['text'] !== "";
     }
 
     /**
      * Get the Message text body
      *
-     * @return mixed
+     * @return string
      */
-    public function getTextBody() {
+    public function getTextBody(): string {
         if (!isset($this->bodies['text'])) {
-            return null;
+            return "";
         }
 
         return $this->bodies['text'];
@@ -377,17 +499,17 @@ class Message {
      * @return bool
      */
     public function hasHTMLBody(): bool {
-        return isset($this->bodies['html']);
+        return isset($this->bodies['html']) && $this->bodies['html'] !== "";
     }
 
     /**
      * Get the Message html body
      *
-     * @return string|null
+     * @return string
      */
-    public function getHTMLBody() {
+    public function getHTMLBody(): string {
         if (!isset($this->bodies['html'])) {
-            return null;
+            return "";
         }
 
         return $this->bodies['html'];
@@ -396,14 +518,18 @@ class Message {
     /**
      * Parse all defined headers
      *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
      * @throws InvalidMessageDateException
      * @throws MessageHeaderFetchingException
+     * @throws ResponseException
      */
-    private function parseHeader() {
+    private function parseHeader(): void {
         $sequence_id = $this->getSequenceId();
-        $headers = $this->client->getConnection()->headers([$sequence_id], "RFC822", $this->sequence === IMAP::ST_UID);
+        $headers = $this->client->getConnection()->headers([$sequence_id], "RFC822", $this->sequence)->validatedData();
         if (!isset($headers[$sequence_id])) {
             throw new MessageHeaderFetchingException("no headers found", 0);
         }
@@ -416,7 +542,7 @@ class Message {
      *
      * @throws InvalidMessageDateException
      */
-    public function parseRawHeader(string $raw_header){
+    public function parseRawHeader(string $raw_header): void {
         $this->header = new Header($raw_header);
     }
 
@@ -424,11 +550,11 @@ class Message {
      * Parse additional raw flags
      * @param array $raw_flags
      */
-    public function parseRawFlags(array $raw_flags) {
+    public function parseRawFlags(array $raw_flags): void {
         $this->flags = FlagCollection::make([]);
 
-        foreach($raw_flags as $flag) {
-            if (strpos($flag, "\\") === 0){
+        foreach ($raw_flags as $flag) {
+            if (str_starts_with($flag, "\\")) {
                 $flag = substr($flag, 1);
             }
             $flag_key = strtolower($flag);
@@ -442,17 +568,21 @@ class Message {
      * Parse additional flags
      *
      * @return void
-     * @throws Exceptions\ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    private function parseFlags() {
+    private function parseFlags(): void {
         $this->client->openFolder($this->folder_path);
         $this->flags = FlagCollection::make([]);
 
         $sequence_id = $this->getSequenceId();
         try {
-            $flags = $this->client->getConnection()->flags([$sequence_id], $this->sequence === IMAP::ST_UID);
+            $flags = $this->client->getConnection()->flags([$sequence_id], $this->sequence)->validatedData();
         } catch (Exceptions\RuntimeException $e) {
             throw new MessageFlagException("flag could not be fetched", 0, $e);
         }
@@ -465,20 +595,24 @@ class Message {
     /**
      * Parse the Message body
      *
-     * @return $this
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\MessageContentFetchingException
+     * @return Message
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
-     * @throws Exceptions\EventNotFoundException
+     * @throws MessageContentFetchingException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function parseBody(): Message {
         $this->client->openFolder($this->folder_path);
 
         $sequence_id = $this->getSequenceId();
         try {
-            $contents = $this->client->getConnection()->content([$sequence_id], "RFC822", $this->sequence === IMAP::ST_UID);
+            $contents = $this->client->getConnection()->content([$sequence_id], "RFC822", $this->sequence)->validatedData();
         } catch (Exceptions\RuntimeException $e) {
             throw new MessageContentFetchingException("failed to fetch content", 0);
         }
@@ -494,19 +628,43 @@ class Message {
     }
 
     /**
+     * Fetches the size for this message.
+     *
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws MessageSizeFetchingException
+     * @throws ResponseException
+     * @throws RuntimeException
+     */
+    private function fetchSize(): void {
+        $sequence_id = $this->getSequenceId();
+        $sizes = $this->client->getConnection()->sizes([$sequence_id], $this->sequence)->validatedData();
+         if (!isset($sizes[$sequence_id])) {
+            throw new MessageSizeFetchingException("sizes did not set an array entry for the supplied sequence_id", 0);
+        }
+        $this->attributes["size"] = $sizes[$sequence_id];
+    }
+
+    /**
      * Handle auto "Seen" flag handling
      *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function peek(){
+    public function peek(): void {
         if ($this->fetch_options == IMAP::FT_PEEK) {
             if ($this->getFlags()->get("seen") == null) {
                 $this->unsetFlag("Seen");
             }
-        }elseif ($this->getFlags()->get("seen") != null) {
+        } elseif ($this->getFlags()->get("seen") == null) {
             $this->setFlag("Seen");
         }
     }
@@ -515,11 +673,15 @@ class Message {
      * Parse a given message body
      * @param string $raw_body
      *
-     * @return $this
-     * @throws Exceptions\ConnectionFailedException
+     * @return Message
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function parseRawBody(string $raw_body): Message {
         $this->structure = new Structure($raw_body, $this->header);
@@ -532,11 +694,15 @@ class Message {
      * Fetch the Message structure
      * @param Structure $structure
      *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    private function fetchStructure(Structure $structure) {
-        $this->client->openFolder($this->folder_path);
+    private function fetchStructure(Structure $structure): void {
+        $this->client?->openFolder($this->folder_path);
 
         foreach ($structure->parts as $part) {
             $this->fetchPart($part);
@@ -547,10 +713,10 @@ class Message {
      * Fetch a given part
      * @param Part $part
      */
-    private function fetchPart(Part $part) {
+    private function fetchPart(Part $part): void {
         if ($part->isAttachment()) {
             $this->fetchAttachment($part);
-        }else{
+        } else {
             $encoding = $this->getEncoding($part);
 
             $content = $this->decodeString($part->content, $part->encoding);
@@ -570,14 +736,27 @@ class Message {
                 $content = $this->convertEncoding($content, $encoding);
             }
 
-            $subtype = strtolower($part->subtype ?? '');
-            $subtype = $subtype == "plain" || $subtype == "" ? "text" : $subtype;
+            $this->addBody($part->subtype ?? '', $content);
+        }
+    }
 
-            if (isset($this->bodies[$subtype])) {
+    /**
+     * Add a body to the message
+     * @param string $subtype
+     * @param string $content
+     *
+     * @return void
+     */
+    protected function addBody(string $subtype, string $content): void {
+        $subtype = strtolower($subtype);
+        $subtype = $subtype == "plain" || $subtype == "" ? "text" : $subtype;
+
+        if (isset($this->bodies[$subtype]) && $this->bodies[$subtype] !== null && $this->bodies[$subtype] !== "") {
+            if ($content !== "") {
                 $this->bodies[$subtype] .= "\n".$content;
-            }else{
-                $this->bodies[$subtype] = $content;
             }
+        } else {
+            $this->bodies[$subtype] = $content;
         }
     }
 
@@ -585,11 +764,11 @@ class Message {
      * Fetch the Message attachment
      * @param Part $part
      */
-    protected function fetchAttachment(Part $part) {
+    protected function fetchAttachment(Part $part): void {
         $oAttachment = new Attachment($this, $part);
 
-        if ($oAttachment->getName() !== null && $oAttachment->getSize() > 0) {
-            if ($oAttachment->getId() !== null) {
+        if ($oAttachment->getSize() > 0) {
+            if ($oAttachment->getId() !== null && $this->attachments->offsetExists($oAttachment->getId())) {
                 $this->attachments->put($oAttachment->getId(), $oAttachment);
             } else {
                 $this->attachments->push($oAttachment);
@@ -601,7 +780,7 @@ class Message {
      * Fail proof setter for $fetch_option
      * @param $option
      *
-     * @return $this
+     * @return Message
      */
     public function setFetchOption($option): Message {
         if (is_long($option) === true) {
@@ -618,9 +797,9 @@ class Message {
      * Set the sequence type
      * @param int|null $sequence
      *
-     * @return $this
+     * @return Message
      */
-    public function setSequence($sequence): Message {
+    public function setSequence(?int $sequence): Message {
         if (is_long($sequence)) {
             $this->sequence = $sequence;
         } elseif (is_null($sequence)) {
@@ -635,7 +814,7 @@ class Message {
      * Fail proof setter for $fetch_body
      * @param $option
      *
-     * @return $this
+     * @return Message
      */
     public function setFetchBodyOption($option): Message {
         if (is_bool($option)) {
@@ -652,7 +831,7 @@ class Message {
      * Fail proof setter for $fetch_flags
      * @param $option
      *
-     * @return $this
+     * @return Message
      */
     public function setFetchFlagsOption($option): Message {
         if (is_bool($option)) {
@@ -699,7 +878,7 @@ class Message {
      *
      * @return mixed|string
      */
-    public function convertEncoding($str, string $from = "ISO-8859-2", string $to = "UTF-8") {
+    public function convertEncoding($str, string $from = "ISO-8859-2", string $to = "UTF-8"): mixed {
 
         $from = EncodingAliases::get($from);
         $to = EncodingAliases::get($to);
@@ -723,8 +902,12 @@ class Message {
             return $str;
         }
 
-        if (function_exists('iconv') && $from != 'UTF-7' && $to != 'UTF-7') {
-            return @iconv($from, $to.'//IGNORE', $str);
+        if (function_exists('iconv') && !EncodingAliases::isUtf7($from) && !EncodingAliases::isUtf7($to)) {
+            try {
+                return iconv($from, $to.'//IGNORE', $str);
+            } catch (\Exception $e) {
+                return @iconv($from, $to, $str);
+            }
         } else {
             if (!$from) {
                 return mb_convert_encoding($str, $to);
@@ -735,21 +918,21 @@ class Message {
 
     /**
      * Get the encoding of a given abject
-     * @param string|object $structure
+     * @param object|string $structure
      *
-     * @return string|null
+     * @return string
      */
-    public function getEncoding($structure): string {
+    public function getEncoding(object|string $structure): string {
         if (property_exists($structure, 'parameters')) {
             foreach ($structure->parameters as $parameter) {
                 if (strtolower($parameter->attribute) == "charset") {
                     return EncodingAliases::get($parameter->value, "ISO-8859-2");
                 }
             }
-        }elseif (property_exists($structure, 'charset')){
+        } elseif (property_exists($structure, 'charset')) {
             return EncodingAliases::get($structure->charset, "ISO-8859-2");
-        }elseif (is_string($structure) === true){
-            return mb_detect_encoding($structure);
+        } elseif (is_string($structure) === true) {
+            return EncodingAliases::detectEncoding($structure);
         }
 
         return 'UTF-8';
@@ -758,12 +941,16 @@ class Message {
     /**
      * Get the messages folder
      *
-     * @return mixed
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @return ?Folder
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function getFolder(){
+    public function getFolder(): ?Folder {
         return $this->client->getFolderByPath($this->folder_path);
     }
 
@@ -773,19 +960,23 @@ class Message {
      * @param MessageCollection|null $thread
      * @param Folder|null $folder
      *
-     * @return MessageCollection|null
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\FolderFetchingException
-     * @throws Exceptions\GetMessagesFailedException
-     * @throws Exceptions\RuntimeException
+     * @return MessageCollection
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws FolderFetchingException
+     * @throws GetMessagesFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function thread(Folder $sent_folder = null, MessageCollection &$thread = null, Folder $folder = null): MessageCollection {
         $thread = $thread ?: MessageCollection::make([]);
-        $folder = $folder ?:  $this->getFolder();
+        $folder = $folder ?: $this->getFolder();
         $sent_folder = $sent_folder ?: $this->client->getFolderByPath(ClientManager::get("options.common_folders.sent", "INBOX/Sent"));
 
         /** @var Message $message */
-        foreach($thread as $message) {
+        foreach ($thread as $message) {
             if ($message->message_id->first() == $this->message_id->first()) {
                 return $thread;
             }
@@ -795,11 +986,9 @@ class Message {
         $this->fetchThreadByInReplyTo($thread, $this->message_id, $folder, $folder, $sent_folder);
         $this->fetchThreadByInReplyTo($thread, $this->message_id, $sent_folder, $folder, $sent_folder);
 
-        if (is_array($this->in_reply_to)) {
-            foreach($this->in_reply_to as $in_reply_to) {
-                $this->fetchThreadByMessageId($thread, $in_reply_to, $folder, $folder, $sent_folder);
-                $this->fetchThreadByMessageId($thread, $in_reply_to, $sent_folder, $folder, $sent_folder);
-            }
+        foreach ($this->in_reply_to->all() as $in_reply_to) {
+            $this->fetchThreadByMessageId($thread, $in_reply_to, $folder, $folder, $sent_folder);
+            $this->fetchThreadByMessageId($thread, $in_reply_to, $sent_folder, $folder, $sent_folder);
         }
 
         return $thread;
@@ -813,18 +1002,22 @@ class Message {
      * @param Folder $secondary_folder
      * @param Folder $sent_folder
      *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\GetMessagesFailedException
-     * @throws Exceptions\RuntimeException
-     * @throws Exceptions\FolderFetchingException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws FolderFetchingException
+     * @throws GetMessagesFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    protected function fetchThreadByInReplyTo(MessageCollection &$thread, string $in_reply_to, Folder $primary_folder, Folder $secondary_folder, Folder $sent_folder){
+    protected function fetchThreadByInReplyTo(MessageCollection &$thread, string $in_reply_to, Folder $primary_folder, Folder $secondary_folder, Folder $sent_folder): void {
         $primary_folder->query()->inReplyTo($in_reply_to)
-        ->setFetchBody($this->getFetchBodyOption())
-        ->leaveUnread()->get()->each(function($message) use(&$thread, $secondary_folder, $sent_folder){
-            /** @var Message $message */
-            $message->thread($sent_folder, $thread, $secondary_folder);
-        });
+            ->setFetchBody($this->getFetchBodyOption())
+            ->leaveUnread()->get()->each(function($message) use (&$thread, $secondary_folder, $sent_folder) {
+                /** @var Message $message */
+                $message->thread($sent_folder, $thread, $secondary_folder);
+            });
     }
 
     /**
@@ -835,18 +1028,22 @@ class Message {
      * @param Folder $secondary_folder
      * @param Folder $sent_folder
      *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\GetMessagesFailedException
-     * @throws Exceptions\RuntimeException
-     * @throws Exceptions\FolderFetchingException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws GetMessagesFailedException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    protected function fetchThreadByMessageId(MessageCollection &$thread, string $message_id, Folder $primary_folder, Folder $secondary_folder, Folder $sent_folder){
+    protected function fetchThreadByMessageId(MessageCollection &$thread, string $message_id, Folder $primary_folder, Folder $secondary_folder, Folder $sent_folder): void {
         $primary_folder->query()->messageId($message_id)
-        ->setFetchBody($this->getFetchBodyOption())
-        ->leaveUnread()->get()->each(function($message) use(&$thread, $secondary_folder, $sent_folder){
-            /** @var Message $message */
-            $message->thread($sent_folder, $thread, $secondary_folder);
-        });
+            ->setFetchBody($this->getFetchBodyOption())
+            ->leaveUnread()->get()->each(function($message) use (&$thread, $secondary_folder, $sent_folder) {
+                /** @var Message $message */
+                $message->thread($sent_folder, $thread, $secondary_folder);
+            });
     }
 
     /**
@@ -855,28 +1052,35 @@ class Message {
      * @param boolean $expunge
      *
      * @return null|Message
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
-     * @throws MessageHeaderFetchingException
-     * @throws Exceptions\EventNotFoundException
      * @throws MessageFlagException
-     * @throws Exceptions\MessageNotFoundException
+     * @throws MessageHeaderFetchingException
+     * @throws MessageNotFoundException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function copy(string $folder_path, bool $expunge = false) {
+    public function copy(string $folder_path, bool $expunge = false): ?Message {
         $this->client->openFolder($folder_path);
-        $status = $this->client->getConnection()->examineFolder($folder_path);
+        $status = $this->client->getConnection()->examineFolder($folder_path)->validatedData();
 
         if (isset($status["uidnext"])) {
             $next_uid = $status["uidnext"];
+            if ((int)$next_uid <= 0) {
+                return null;
+            }
 
             /** @var Folder $folder */
             $folder = $this->client->getFolderByPath($folder_path);
 
             $this->client->openFolder($this->folder_path);
-            if ($this->client->getConnection()->copyMessage($folder->path, $this->getSequenceId(), null, $this->sequence === IMAP::ST_UID) == true) {
+            if ($this->client->getConnection()->copyMessage($folder->path, $this->getSequenceId(), null, $this->sequence)->validatedData()) {
                 return $this->fetchNewMail($folder, $next_uid, "copied", $expunge);
             }
         }
@@ -890,28 +1094,35 @@ class Message {
      * @param boolean $expunge
      *
      * @return Message|null
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\FolderFetchingException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
-     * @throws MessageHeaderFetchingException
-     * @throws Exceptions\EventNotFoundException
      * @throws MessageFlagException
-     * @throws Exceptions\MessageNotFoundException
+     * @throws MessageHeaderFetchingException
+     * @throws MessageNotFoundException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function move(string $folder_path, bool $expunge = false) {
+    public function move(string $folder_path, bool $expunge = false): ?Message {
         $this->client->openFolder($folder_path);
-        $status = $this->client->getConnection()->examineFolder($folder_path);
+        $status = $this->client->getConnection()->examineFolder($folder_path)->validatedData();
 
         if (isset($status["uidnext"])) {
             $next_uid = $status["uidnext"];
+            if ((int)$next_uid <= 0) {
+                return null;
+            }
 
             /** @var Folder $folder */
             $folder = $this->client->getFolderByPath($folder_path);
 
             $this->client->openFolder($this->folder_path);
-            if ($this->client->getConnection()->moveMessage($folder->path, $this->getSequenceId(), null, $this->sequence === IMAP::ST_UID) == true) {
+            if ($this->client->getConnection()->moveMessage($folder->path, $this->getSequenceId(), null, $this->sequence)->validatedData()) {
                 return $this->fetchNewMail($folder, $next_uid, "moved", $expunge);
             }
         }
@@ -927,24 +1138,28 @@ class Message {
      * @param boolean $expunge
      *
      * @return Message
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
-     * @throws Exceptions\MessageNotFoundException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
      * @throws MessageFlagException
      * @throws MessageHeaderFetchingException
+     * @throws MessageNotFoundException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     protected function fetchNewMail(Folder $folder, int $next_uid, string $event, bool $expunge): Message {
-        if($expunge) $this->client->expunge();
+        if ($expunge) $this->client->expunge();
 
         $this->client->openFolder($folder->path);
 
         if ($this->sequence === IMAP::ST_UID) {
             $sequence_id = $next_uid;
-        }else{
-            $sequence_id = $this->client->getConnection()->getMessageNumber($next_uid);
+        } else {
+            $sequence_id = $this->client->getConnection()->getMessageNumber($next_uid)->validatedData();
         }
 
         $message = $folder->query()->getMessage($sequence_id, null, $this->sequence);
@@ -961,23 +1176,27 @@ class Message {
      * @param boolean $force_move
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
-     * @throws Exceptions\FolderFetchingException
-     * @throws Exceptions\MessageNotFoundException
-     * @throws Exceptions\RuntimeException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws FolderFetchingException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws InvalidMessageDateException
      * @throws MessageContentFetchingException
      * @throws MessageFlagException
      * @throws MessageHeaderFetchingException
+     * @throws MessageNotFoundException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function delete(bool $expunge = true, string $trash_path = null, bool $force_move = false) {
+    public function delete(bool $expunge = true, string $trash_path = null, bool $force_move = false): bool {
         $status = $this->setFlag("Deleted");
-        if($force_move) {
-            $trash_path = $trash_path === null ? $this->config["common_folders"]["trash"]: $trash_path;
-            $status = $this->move($trash_path);
+        if ($force_move) {
+            $trash_path = $trash_path === null ? $this->config["common_folders"]["trash"] : $trash_path;
+            $this->move($trash_path);
         }
-        if($expunge) $this->client->expunge();
+        if ($expunge) $this->client->expunge();
 
         $event = $this->getEvent("message", "deleted");
         $event::dispatch($this);
@@ -990,14 +1209,18 @@ class Message {
      * @param boolean $expunge
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function restore(bool $expunge = true): bool {
         $status = $this->unsetFlag("Deleted");
-        if($expunge) $this->client->expunge();
+        if ($expunge) $this->client->expunge();
 
         $event = $this->getEvent("message", "restored");
         $event::dispatch($this);
@@ -1007,20 +1230,24 @@ class Message {
 
     /**
      * Set a given flag
-     * @param string|array $flag
+     * @param array|string $flag
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\EventNotFoundException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function setFlag($flag): bool {
+    public function setFlag(array|string $flag): bool {
         $this->client->openFolder($this->folder_path);
-        $flag = "\\".trim(is_array($flag) ? implode(" \\", $flag) : $flag);
+        $flag = "\\" . trim(is_array($flag) ? implode(" \\", $flag) : $flag);
         $sequence_id = $this->getSequenceId();
         try {
-            $status = $this->client->getConnection()->store([$flag], $sequence_id, $sequence_id, "+", true, $this->sequence === IMAP::ST_UID);
+            $status = $this->client->getConnection()->store([$flag], $sequence_id, $sequence_id, "+", true, $this->sequence)->validatedData();
         } catch (Exceptions\RuntimeException $e) {
             throw new MessageFlagException("flag could not be set", 0, $e);
         }
@@ -1034,21 +1261,25 @@ class Message {
 
     /**
      * Unset a given flag
-     * @param string|array $flag
+     * @param array|string $flag
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function unsetFlag($flag): bool {
+    public function unsetFlag(array|string $flag): bool {
         $this->client->openFolder($this->folder_path);
 
-        $flag = "\\".trim(is_array($flag) ? implode(" \\", $flag) : $flag);
+        $flag = "\\" . trim(is_array($flag) ? implode(" \\", $flag) : $flag);
         $sequence_id = $this->getSequenceId();
         try {
-            $status = $this->client->getConnection()->store([$flag], $sequence_id, $sequence_id, "-", true, $this->sequence === IMAP::ST_UID);
+            $status = $this->client->getConnection()->store([$flag], $sequence_id, $sequence_id, "-", true, $this->sequence)->validatedData();
         } catch (Exceptions\RuntimeException $e) {
             throw new MessageFlagException("flag could not be removed", 0, $e);
         }
@@ -1062,29 +1293,37 @@ class Message {
 
     /**
      * Set a given flag
-     * @param string|array $flag
+     * @param array|string $flag
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\EventNotFoundException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function addFlag($flag): bool {
+    public function addFlag(array|string $flag): bool {
         return $this->setFlag($flag);
     }
 
     /**
      * Unset a given flag
-     * @param string|array $flag
+     * @param array|string $flag
      *
      * @return bool
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\EventNotFoundException
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws EventNotFoundException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws MessageFlagException
-     * @throws Exceptions\RuntimeException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
-    public function removeFlag($flag): bool {
+    public function removeFlag(array|string $flag): bool {
         return $this->unsetFlag($flag);
     }
 
@@ -1119,13 +1358,9 @@ class Message {
      * Get the raw body
      *
      * @return string
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\RuntimeException
      */
-    public function getRawBody() {
-        if ($this->raw_body === null) {
-            $this->client->openFolder($this->folder_path);
-
+    public function getRawBody(): string {
+        if ($this->raw_body === "") {
             $this->raw_body = $this->structure->raw;
         }
 
@@ -1135,27 +1370,27 @@ class Message {
     /**
      * Get the message header
      *
-     * @return Header
+     * @return ?Header
      */
-    public function getHeader() {
+    public function getHeader(): ?Header {
         return $this->header;
     }
 
     /**
      * Get the current client
      *
-     * @return Client
+     * @return ?Client
      */
-    public function getClient(): Client {
+    public function getClient(): ?Client {
         return $this->client;
     }
 
     /**
      * Get the used fetch option
      *
-     * @return integer
+     * @return ?integer
      */
-    public function getFetchOptions() {
+    public function getFetchOptions(): ?int {
         return $this->fetch_options;
     }
 
@@ -1164,7 +1399,7 @@ class Message {
      *
      * @return boolean
      */
-    public function getFetchBodyOption() {
+    public function getFetchBodyOption(): bool {
         return $this->fetch_body;
     }
 
@@ -1173,7 +1408,7 @@ class Message {
      *
      * @return boolean
      */
-    public function getFetchFlagsOption() {
+    public function getFetchFlagsOption(): bool {
         return $this->fetch_flags;
     }
 
@@ -1205,18 +1440,29 @@ class Message {
     }
 
     /**
+     * Check if a flag is set
+     *
+     * @param string $flag
+     * @return boolean
+     */
+    public function hasFlag(string $flag): bool {
+        $flag = str_replace("\\", "", strtolower($flag));
+        return $this->getFlags()->has($flag);
+    }
+
+    /**
      * Get the fetched structure
      *
      * @return Structure|null
      */
-    public function getStructure(){
+    public function getStructure(): ?Structure {
         return $this->structure;
     }
 
     /**
-     * Check if a message matches an other by comparing basic attributes
+     * Check if a message matches another by comparing basic attributes
      *
-     * @param  null|Message $message
+     * @param null|Message $message
      * @return boolean
      */
     public function is(Message $message = null): bool {
@@ -1227,7 +1473,7 @@ class Message {
         return $this->uid == $message->uid
             && $this->message_id->first() == $message->message_id->first()
             && $this->subject->first() == $message->subject->first()
-            && $this->date->toDate()->eq($message->date);
+            && $this->date->toDate()->eq($message->date->toDate());
     }
 
     /**
@@ -1243,10 +1489,10 @@ class Message {
      * Set the message mask
      * @param $mask
      *
-     * @return $this
+     * @return Message
      */
     public function setMask($mask): Message {
-        if(class_exists($mask)){
+        if (class_exists($mask)) {
             $this->mask = $mask;
         }
 
@@ -1264,18 +1510,18 @@ class Message {
 
     /**
      * Get a masked instance by providing a mask name
-     * @param string|mixed $mask
+     * @param mixed|null $mask
      *
      * @return mixed
      * @throws MaskNotFoundException
      */
-    public function mask($mask = null){
+    public function mask(mixed $mask = null): mixed {
         $mask = $mask !== null ? $mask : $this->mask;
-        if(class_exists($mask)){
+        if (class_exists($mask)) {
             return new $mask($this);
         }
 
-        throw new MaskNotFoundException("Unknown mask provided: ".$mask);
+        throw new MaskNotFoundException("Unknown mask provided: " . $mask);
     }
 
     /**
@@ -1291,7 +1537,7 @@ class Message {
      * Set the message path aka folder path
      * @param $folder_path
      *
-     * @return $this
+     * @return Message
      */
     public function setFolderPath($folder_path): Message {
         $this->folder_path = $folder_path;
@@ -1301,21 +1547,30 @@ class Message {
 
     /**
      * Set the config
-     * @param $config
+     * @param array $config
      *
-     * @return $this
+     * @return Message
      */
-    public function setConfig($config): Message {
+    public function setConfig(array $config): Message {
         $this->config = $config;
 
         return $this;
     }
 
     /**
+     * Get the config
+     *
+     * @return array
+     */
+    public function getConfig(): array {
+        return $this->config;
+    }
+
+    /**
      * Set the available flags
      * @param $available_flags
      *
-     * @return $this
+     * @return Message
      */
     public function setAvailableFlags($available_flags): Message {
         $this->available_flags = $available_flags;
@@ -1324,10 +1579,19 @@ class Message {
     }
 
     /**
+     * Get the available flags
+     *
+     * @return array
+     */
+    public function getAvailableFlags(): array {
+        return $this->available_flags;
+    }
+
+    /**
      * Set the attachment collection
      * @param $attachments
      *
-     * @return $this
+     * @return Message
      */
     public function setAttachments($attachments): Message {
         $this->attachments = $attachments;
@@ -1339,7 +1603,7 @@ class Message {
      * Set the flag collection
      * @param $flags
      *
-     * @return $this
+     * @return Message
      */
     public function setFlags($flags): Message {
         $this->flags = $flags;
@@ -1351,13 +1615,17 @@ class Message {
      * Set the client
      * @param $client
      *
-     * @return $this
-     * @throws Exceptions\RuntimeException
-     * @throws Exceptions\ConnectionFailedException
+     * @return Message
+     * @throws AuthFailedException
+     * @throws ConnectionFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
+     * @throws ResponseException
      */
     public function setClient($client): Message {
         $this->client = $client;
-        $this->client->openFolder($this->folder_path);
+        $this->client?->openFolder($this->folder_path);
 
         return $this;
     }
@@ -1366,13 +1634,11 @@ class Message {
      * Set the message number
      * @param int $uid
      *
-     * @return $this
-     * @throws Exceptions\MessageNotFoundException
-     * @throws Exceptions\ConnectionFailedException
+     * @return Message
      */
     public function setUid(int $uid): Message {
         $this->uid = $uid;
-        $this->msgn = $this->client->getConnection()->getMessageNumber($this->uid);
+        $this->msgn = null;
         $this->msglist = null;
 
         return $this;
@@ -1383,14 +1649,12 @@ class Message {
      * @param int $msgn
      * @param int|null $msglist
      *
-     * @return $this
-     * @throws Exceptions\MessageNotFoundException
-     * @throws Exceptions\ConnectionFailedException
+     * @return Message
      */
     public function setMsgn(int $msgn, int $msglist = null): Message {
         $this->msgn = $msgn;
         $this->msglist = $msglist;
-        $this->uid = $this->client->getConnection()->getUid($this->msgn);
+        $this->uid = null;
 
         return $this;
     }
@@ -1405,7 +1669,7 @@ class Message {
     }
 
     /**
-     * Set the sequence type
+     * Get the current sequence id (either a UID or a message number!)
      *
      * @return int
      */
@@ -1417,16 +1681,23 @@ class Message {
      * Set the sequence id
      * @param $uid
      * @param int|null $msglist
-     *
-     * @throws Exceptions\ConnectionFailedException
-     * @throws Exceptions\MessageNotFoundException
      */
-    public function setSequenceId($uid, int $msglist = null){
+    public function setSequenceId($uid, int $msglist = null): void {
         if ($this->getSequence() === IMAP::ST_UID) {
             $this->setUid($uid);
             $this->setMsglist($msglist);
-        }else{
+        } else {
             $this->setMsgn($uid, $msglist);
         }
+    }
+
+    /**
+     * Safe the entire message in a file
+     * @param $filename
+     *
+     * @return bool|int
+     */
+    public function save($filename): bool|int {
+        return file_put_contents($filename, $this->header->raw."\r\n\r\n".$this->structure->raw);
     }
 }
