@@ -27,117 +27,117 @@ class Part {
      *
      * @var string $raw
      */
-    public $raw = "";
+    public string $raw = "";
 
     /**
      * Part type
      *
      * @var int $type
      */
-    public $type = IMAP::MESSAGE_TYPE_TEXT;
+    public int $type = IMAP::MESSAGE_TYPE_TEXT;
 
     /**
      * Part content
      *
      * @var string $content
      */
-    public $content = "";
+    public string $content = "";
 
     /**
      * Part subtype
      *
-     * @var string $subtype
+     * @var ?string $subtype
      */
-    public $subtype = null;
+    public ?string $subtype = null;
 
     /**
      * Part charset - if available
      *
      * @var string $charset
      */
-    public $charset = "utf-8";
+    public string $charset = "utf-8";
 
     /**
      * Part encoding method
      *
      * @var int $encoding
      */
-    public $encoding = IMAP::MESSAGE_ENC_OTHER;
+    public int $encoding = IMAP::MESSAGE_ENC_OTHER;
 
     /**
      * Alias to check if the part is an attachment
      *
      * @var boolean $ifdisposition
      */
-    public $ifdisposition = false;
+    public bool $ifdisposition = false;
 
     /**
      * Indicates if the part is an attachment
      *
-     * @var string $disposition
+     * @var ?string $disposition
      */
-    public $disposition = null;
+    public ?string $disposition = null;
 
     /**
      * Alias to check if the part has a description
      *
      * @var boolean $ifdescription
      */
-    public $ifdescription = false;
+    public bool $ifdescription = false;
 
     /**
      * Part description if available
      *
-     * @var string $description
+     * @var ?string $description
      */
-    public $description = null;
+    public ?string $description = null;
 
     /**
      * Part filename if available
      *
-     * @var string $filename
+     * @var ?string $filename
      */
-    public $filename = null;
+    public ?string $filename = null;
 
     /**
      * Part name if available
      *
-     * @var string $name
+     * @var ?string $name
      */
-    public $name = null;
+    public ?string $name = null;
 
     /**
      * Part id if available
      *
-     * @var string $id
+     * @var ?string $id
      */
-    public $id = null;
+    public ?string $id = null;
 
     /**
      * The part number of the current part
      *
      * @var integer $part_number
      */
-    public $part_number = 0;
+    public int $part_number = 0;
 
     /**
      * Part length in bytes
      *
      * @var integer $bytes
      */
-    public $bytes = null;
+    public int $bytes;
 
     /**
      * Part content type
      *
      * @var string|null $content_type
      */
-    public $content_type = null;
+    public ?string $content_type = null;
 
     /**
-     * @var Header $header
+     * @var ?Header $header
      */
-    private $header = null;
+    private ?Header $header;
 
     /**
      * Part constructor.
@@ -159,7 +159,7 @@ class Part {
      *
      * @throws InvalidMessageDateException
      */
-    protected function parse(){
+    protected function parse(): void {
         if ($this->header === null) {
             $body = $this->findHeaders();
         }else{
@@ -170,32 +170,28 @@ class Part {
         $this->parseDescription();
         $this->parseEncoding();
 
-        $this->charset = $this->header->get("charset");
+        $this->charset = $this->header->get("charset")->first();
         $this->name = $this->header->get("name");
         $this->filename = $this->header->get("filename");
 
-        if(!empty($this->header->get("id"))) {
+        if($this->header->get("id")->exist()) {
             $this->id = $this->header->get("id");
-        } else if(!empty($this->header->get("x_attachment_id"))){
+        } else if($this->header->get("x_attachment_id")->exist()){
             $this->id = $this->header->get("x_attachment_id");
-        } else if(!empty($this->header->get("content_id"))){
+        } else if($this->header->get("content_id")->exist()){
             $this->id = strtr($this->header->get("content_id"), [
                 '<' => '',
                 '>' => ''
             ]);
         }
 
-        $content_types = $this->header->get("content_type");
+        $content_types = $this->header->get("content_type")->all();
         if(!empty($content_types)){
             $this->subtype = $this->parseSubtype($content_types);
-            $content_type = $content_types;
-            if (is_array($content_types)) {
-                $content_type = $content_types[0];
-            }
+            $content_type = $content_types[0];
             $parts = explode(';', $content_type);
             $this->content_type = trim($parts[0]);
         }
-
 
         $this->content = trim(rtrim($body));
         $this->bytes = strlen($this->content);
@@ -224,9 +220,9 @@ class Part {
      * Try to parse the subtype if any is present
      * @param $content_type
      *
-     * @return string
+     * @return ?string
      */
-    private function parseSubtype($content_type){
+    private function parseSubtype($content_type): ?string {
         if (is_array($content_type)) {
             foreach ($content_type as $part){
                 if ((strpos($part, "/")) !== false){
@@ -236,7 +232,7 @@ class Part {
             return null;
         }
         if (($pos = strpos($content_type, "/")) !== false){
-            return substr($content_type, $pos + 1);
+            return substr(explode(";", $content_type)[0], $pos + 1);
         }
         return null;
     }
@@ -244,20 +240,20 @@ class Part {
     /**
      * Try to parse the disposition if any is present
      */
-    private function parseDisposition(){
-        $content_disposition = $this->header->get("content_disposition");
-        if($content_disposition !== null) {
+    private function parseDisposition(): void {
+        $content_disposition = $this->header->get("content_disposition")->first();
+        if($content_disposition) {
             $this->ifdisposition = true;
-            $this->disposition = (is_array($content_disposition)) ? implode(' ', $content_disposition) : $content_disposition;
+            $this->disposition = (is_array($content_disposition)) ? implode(' ', $content_disposition) : explode(";", $content_disposition)[0];
         }
     }
 
     /**
      * Try to parse the description if any is present
      */
-    private function parseDescription(){
-        $content_description = $this->header->get("content_description");
-        if($content_description !== null) {
+    private function parseDescription(): void {
+        $content_description = $this->header->get("content_description")->first();
+        if($content_description) {
             $this->ifdescription = true;
             $this->description = $content_description;
         }
@@ -266,30 +262,17 @@ class Part {
     /**
      * Try to parse the encoding if any is present
      */
-    private function parseEncoding(){
-        $encoding = $this->header->get("content_transfer_encoding");
-        if($encoding !== null) {
-            switch (strtolower($encoding)) {
-                case "quoted-printable":
-                    $this->encoding = IMAP::MESSAGE_ENC_QUOTED_PRINTABLE;
-                    break;
-                case "base64":
-                    $this->encoding = IMAP::MESSAGE_ENC_BASE64;
-                    break;
-                case "7bit":
-                    $this->encoding = IMAP::MESSAGE_ENC_7BIT;
-                    break;
-                case "8bit":
-                    $this->encoding = IMAP::MESSAGE_ENC_8BIT;
-                    break;
-                case "binary":
-                    $this->encoding = IMAP::MESSAGE_ENC_BINARY;
-                    break;
-                default:
-                    $this->encoding = IMAP::MESSAGE_ENC_OTHER;
-                    break;
-
-            }
+    private function parseEncoding(): void {
+        $encoding = $this->header->get("content_transfer_encoding")->first();
+        if($encoding) {
+            $this->encoding = match (strtolower($encoding)) {
+                "quoted-printable" => IMAP::MESSAGE_ENC_QUOTED_PRINTABLE,
+                "base64" => IMAP::MESSAGE_ENC_BASE64,
+                "7bit" => IMAP::MESSAGE_ENC_7BIT,
+                "8bit" => IMAP::MESSAGE_ENC_8BIT,
+                "binary" => IMAP::MESSAGE_ENC_BINARY,
+                default => IMAP::MESSAGE_ENC_OTHER,
+            };
         }
     }
 
@@ -306,7 +289,20 @@ class Part {
                 return false;
             }
         }
+
+        if ($this->disposition === "inline" && $this->filename == null && $this->name == null && !$this->header->has("content_id")) {
+            return false;
+        }
         return true;
+    }
+
+    /**
+     * Get the part header
+     *
+     * @return Header|null
+     */
+    public function getHeader(): ?Header {
+        return $this->header;
     }
 
 }
