@@ -28,7 +28,7 @@ class PaginatedCollection extends Collection {
      *
      * @var int $total
      */
-    protected $total;
+    protected int $total = 0;
 
     /**
      * Paginate the current collection.
@@ -39,7 +39,7 @@ class PaginatedCollection extends Collection {
      *
      * @return LengthAwarePaginator
      */
-    public function paginate(int $per_page = 15, $page = null, string $page_name = 'page', bool $prepaginated = false): LengthAwarePaginator {
+    public function paginate(int $per_page = 15, ?int $page = null, string $page_name = 'page', bool $prepaginated = false): LengthAwarePaginator {
         $page = $page ?: Paginator::resolveCurrentPage($page_name);
 
         $total = $this->total ?: $this->count();
@@ -57,12 +57,12 @@ class PaginatedCollection extends Collection {
      * @param array $items
      * @param int $total
      * @param int $per_page
-     * @param  int|null $current_page
+     * @param int|null $current_page
      * @param  array    $options
      *
      * @return LengthAwarePaginator
      */
-    protected function paginator(array $items, int $total, int $per_page, $current_page, array $options): LengthAwarePaginator {
+    protected function paginator(array $items, int $total, int $per_page, ?int $current_page, array $options): LengthAwarePaginator {
         return new LengthAwarePaginator($items, $total, $per_page, $current_page, $options);
     }
 
@@ -72,7 +72,7 @@ class PaginatedCollection extends Collection {
      *
      * @return int|null
      */
-    public function total($total = null) {
+    public function total($total = null): ?int {
         if($total === null) {
             return $this->total;
         }
