@@ -16,7 +16,7 @@ class CustomAttachmentMask extends \Webklex\PHPIMAP\Support\Masks\AttachmentMask
      * New custom method which can be called through a mask
      * @return string
      */
-    public function token(){
+    public function token(): string {
         return implode('-', [$this->id, $this->getMessage()->getUid(), $this->name]);
     }
 
@@ -24,7 +24,7 @@ class CustomAttachmentMask extends \Webklex\PHPIMAP\Support\Masks\AttachmentMask
      * Custom attachment saving method
      * @return bool
      */
-    public function custom_save() {
+    public function custom_save(): bool {
         $path = "foo".DIRECTORY_SEPARATOR."bar".DIRECTORY_SEPARATOR;
         $filename = $this->token();
 
