@@ -13,9 +13,10 @@
 namespace Webklex\PHPIMAP\Connection\Protocols;
 
 use ErrorException;
-use Webklex\PHPIMAP\Client;
 use Webklex\PHPIMAP\Exceptions\AuthFailedException;
 use Webklex\PHPIMAP\Exceptions\ConnectionFailedException;
+use Webklex\PHPIMAP\Exceptions\ImapBadRequestException;
+use Webklex\PHPIMAP\Exceptions\ImapServerErrorException;
 use Webklex\PHPIMAP\Exceptions\InvalidMessageDateException;
 use Webklex\PHPIMAP\Exceptions\MessageNotFoundException;
 use Webklex\PHPIMAP\Exceptions\RuntimeException;
@@ -30,6 +31,10 @@ interface ProtocolInterface {
 
     /**
      * Public destructor
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
      */
     public function __destruct();
 
@@ -42,34 +47,42 @@ interface ProtocolInterface {
      * @throws ConnectionFailedException
      * @throws RuntimeException
      */
-    public function connect(string $host, $port = null);
+    public function connect(string $host, ?int $port = null);
 
     /**
      * Login to a new session.
      *
      * @param string $user username
      * @param string $password password
-     * @return bool success
+     *
+     * @return Response
+     *
      * @throws AuthFailedException
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      */
-    public function login(string $user, string $password): bool;
+    public function login(string $user, string $password): Response;
 
     /**
      * Authenticate your current session.
      * @param string $user username
      * @param string $token access token
      *
-     * @return bool|mixed
+     * @return Response
      * @throws AuthFailedException
      */
-    public function authenticate(string $user, string $token);
+    public function authenticate(string $user, string $token): Response;
 
     /**
      * Logout of the current server session
      *
-     * @return bool success
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
+     * @throws RuntimeException
      */
-    public function logout(): bool;
+    public function logout(): Response;
 
     /**
      * Check if the current session is connected
@@ -81,95 +94,106 @@ interface ProtocolInterface {
     /**
      * Get an array of available capabilities
      *
-     * @return array list of capabilities
+     * @return Response containing a list of capabilities
      * @throws RuntimeException
      */
-    public function getCapabilities(): array;
+    public function getCapabilities(): Response;
 
     /**
      * Change the current folder
-     *
      * @param string $folder change to this folder
-     * @return bool|array see examineOrSelect()
+     *
+     * @return Response see examineOrSelect()
      * @throws RuntimeException
      */
-    public function selectFolder(string $folder = 'INBOX');
+    public function selectFolder(string $folder = 'INBOX'): Response;
 
     /**
      * Examine a given folder
-     *
      * @param string $folder
-     * @return bool|array
+     *
+     * @return Response
      * @throws RuntimeException
      */
-    public function examineFolder(string $folder = 'INBOX');
+    public function examineFolder(string $folder = 'INBOX'): Response;
 
     /**
      * Fetch message headers
-     * @param array|int $uids
+     * @param int|array $uids
      * @param string $rfc
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array
+     * @return Response
      * @throws RuntimeException
      */
-    public function content($uids, string $rfc = "RFC822", $uid = IMAP::ST_UID): array;
+    public function content(int|array $uids, string $rfc = "RFC822", int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Fetch message headers
-     * @param array|int $uids
+     * @param int|array $uids
      * @param string $rfc
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array
+     * @return Response
      * @throws RuntimeException
      */
-    public function headers($uids, string $rfc = "RFC822", $uid = IMAP::ST_UID): array;
+    public function headers(int|array $uids, string $rfc = "RFC822", int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Fetch message flags
-     * @param array|int $uids
+     * @param int|array $uids
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array
+     * @return Response
      * @throws RuntimeException
      */
-    public function flags($uids, $uid = IMAP::ST_UID): array;
+    public function flags(int|array $uids, int|string $uid = IMAP::ST_UID): Response;
+    
+    /**
+     * Fetch message sizes
+     * @param int|array $uids
+     * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
+     * message numbers instead.
+     *
+     * @return Response
+     * @throws RuntimeException
+     */
+    public function sizes(int|array $uids, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Get uid for a given id
      * @param int|null $id message number
      *
-     * @return array|string message number for given message or all messages as array
+     * @return Response containing a message number for given message or all messages as array
      * @throws MessageNotFoundException
      */
-    public function getUid($id = null);
+    public function getUid(?int $id = null): Response;
 
     /**
      * Get a message number for a uid
      * @param string $id uid
      *
-     * @return int message number
+     * @return Response containing the message number
      * @throws MessageNotFoundException
      */
-    public function getMessageNumber(string $id): int;
+    public function getMessageNumber(string $id): Response;
 
     /**
      * Get a list of available folders
      * @param string $reference mailbox reference for list
      * @param string $folder mailbox / folder name match with wildcards
      *
-     * @return array mailboxes that matched $folder as array(globalName => array('delim' => .., 'flags' => ..))
+     * @return Response containing mailboxes that matched $folder as array(globalName => array('delim' => .., 'flags' => ..))
      * @throws RuntimeException
      */
-    public function folders(string $reference = '', string $folder = '*'): array;
+    public function folders(string $reference = '', string $folder = '*'): Response;
 
     /**
      * Set message flags
-     * @param array $flags flags to set, add or remove
+     * @param array|string $flags flags to set, add or remove
      * @param int $from message for items or start message if $to !== null
      * @param int|null $to if null only one message ($from) is fetched, else it's the
      *                             last message, INF means last message available
@@ -177,12 +201,12 @@ interface ProtocolInterface {
      * @param bool $silent if false the return values are the new flags for the wanted messages
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
-     * @param null|string $item command used to store a flag
+     * @param string|null $item command used to store a flag
      *
-     * @return bool|array new flags if $silent is false, else true or false depending on success
+     * @return Response containing the new flags if $silent is false, else true or false depending on success
      * @throws RuntimeException
      */
-    public function store(array $flags, int $from, $to = null, $mode = null, bool $silent = true, $uid = IMAP::ST_UID, $item = null);
+    public function store(array|string $flags, int $from, ?int $to = null, ?string $mode = null, bool $silent = true, int|string $uid = IMAP::ST_UID, ?string $item = null): Response;
 
     /**
      * Append a new message to given folder
@@ -191,10 +215,10 @@ interface ProtocolInterface {
      * @param array|null $flags flags for new message
      * @param string|null $date date for new message
      *
-     * @return bool success
+     * @return Response
      * @throws RuntimeException
      */
-    public function appendMessage(string $folder, string $message, $flags = null, $date = null): bool;
+    public function appendMessage(string $folder, string $message, ?array $flags = null, ?string $date = null): Response;
 
     /**
      * Copy message set from current folder to other folder
@@ -206,10 +230,10 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return bool success
+     * @return Response
      * @throws RuntimeException
      */
-    public function copyMessage(string $folder, $from, $to = null, $uid = IMAP::ST_UID): bool;
+    public function copyMessage(string $folder, $from, ?int $to = null, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Copy multiple messages to the target folder
@@ -218,10 +242,10 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array|bool Tokens if operation successful, false if an error occurred
+     * @return Response Tokens if operation successful, false if an error occurred
      * @throws RuntimeException
      */
-    public function copyManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID);
+    public function copyManyMessages(array $messages, string $folder, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Move a message set from current folder to another folder
@@ -232,9 +256,9 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return bool success
+     * @return Response
      */
-    public function moveMessage(string $folder, $from, $to = null, $uid = IMAP::ST_UID): bool;
+    public function moveMessage(string $folder, $from, ?int $to = null, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Move multiple messages to the target folder
@@ -244,67 +268,76 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array|bool Tokens if operation successful, false if an error occurred
+     * @return Response Tokens if operation successful, false if an error occurred
      * @throws RuntimeException
      */
-    public function moveManyMessages(array $messages, string $folder, $uid = IMAP::ST_UID);
+    public function moveManyMessages(array $messages, string $folder, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Exchange identification information
      * Ref.: https://datatracker.ietf.org/doc/html/rfc2971
      *
      * @param null $ids
-     * @return array|bool|void|null
+     * @return Response
      *
      * @throws RuntimeException
      */
-    public function ID($ids = null);
+    public function ID($ids = null): Response;
 
     /**
      * Create a new folder
      *
      * @param string $folder folder name
-     * @return bool success
+     * @return Response
      * @throws RuntimeException
      */
-    public function createFolder(string $folder): bool;
+    public function createFolder(string $folder): Response;
 
     /**
      * Rename an existing folder
      *
      * @param string $old old name
      * @param string $new new name
-     * @return bool success
+     * @return Response
      * @throws RuntimeException
      */
-    public function renameFolder(string $old, string $new): bool;
+    public function renameFolder(string $old, string $new): Response;
 
     /**
      * Delete a folder
      *
      * @param string $folder folder name
-     * @return bool success
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws RuntimeException
      */
-    public function deleteFolder(string $folder): bool;
+    public function deleteFolder(string $folder): Response;
 
     /**
      * Subscribe to a folder
      *
      * @param string $folder folder name
-     * @return bool success
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws RuntimeException
      */
-    public function subscribeFolder(string $folder): bool;
+    public function subscribeFolder(string $folder): Response;
 
     /**
      * Unsubscribe from a folder
-     * @param string $folder folder name
      *
-     * @return bool success
+     * @param string $folder folder name
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws RuntimeException
      */
-    public function unsubscribeFolder(string $folder): bool;
+    public function unsubscribeFolder(string $folder): Response;
 
     /**
      * Send idle command
@@ -322,37 +355,43 @@ interface ProtocolInterface {
     /**
      * Apply session saved changes to the server
      *
-     * @return bool success
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws RuntimeException
      */
-    public function expunge(): bool;
+    public function expunge(): Response;
 
     /**
      * Retrieve the quota level settings, and usage statics per mailbox
      * @param $username
      *
-     * @return array
+     * @return Response
      * @throws RuntimeException
      */
-    public function getQuota($username): array;
+    public function getQuota($username): Response;
 
     /**
      * Retrieve the quota settings per user
      *
      * @param string $quota_root
      *
-     * @return array
+     * @return Response
      * @throws ConnectionFailedException
      */
-    public function getQuotaRoot(string $quota_root = 'INBOX'): array;
+    public function getQuotaRoot(string $quota_root = 'INBOX'): Response;
 
     /**
      * Send noop command
      *
-     * @return bool success
+     * @return Response
+     *
+     * @throws ImapBadRequestException
+     * @throws ImapServerErrorException
      * @throws RuntimeException
      */
-    public function noop(): bool;
+    public function noop(): Response;
 
     /**
      * Do a search request
@@ -361,10 +400,10 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array message ids
+     * @return Response containing the message ids
      * @throws RuntimeException
      */
-    public function search(array $params, $uid = IMAP::ST_UID): array;
+    public function search(array $params, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Get a message overview
@@ -372,12 +411,12 @@ interface ProtocolInterface {
      * @param int|string $uid set to IMAP::ST_UID or any string representing the UID - set to IMAP::ST_MSGN to use
      * message numbers instead.
      *
-     * @return array
+     * @return Response
      * @throws RuntimeException
      * @throws MessageNotFoundException
      * @throws InvalidMessageDateException
      */
-    public function overview(string $sequence, $uid = IMAP::ST_UID): array;
+    public function overview(string $sequence, int|string $uid = IMAP::ST_UID): Response;
 
     /**
      * Enable the debug mode
@@ -404,5 +443,5 @@ interface ProtocolInterface {
      *
      * @param array|null $uids
      */
-    public function setUidCache($uids);
+    public function setUidCache(?array $uids);
 }
